[roll] Roll fuchsia [fxfs] Change compaction strategy
Prior to this change, we would always compact all layers but that had
the potential to exhaust the journal, so this change introduces a limit
for object store compactions so that we will only compact some of the
layers to keep below a given limit.

This fixes the DirectoryTest.Large/Fxfs test which got broken with a
previous change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534021
Original-Revision: a16c5265d1e2428a45dbb4ce1d161890f4eee557
GitOrigin-RevId: fd759942c47b837925af8c1a63990cd47d6c1d35
Change-Id: Ia17dde1331c75c7a7e8e366265ff9d030289e312