[roll] Roll fuchsia [fxfs] Sequence Items in the tree
Adds a monotonically increasing sequence number to Item which can be
used for a number of purposes:

- Filtering out items which were inserted while a graveyard purge was
  occurring, so that we only purge items that were already in the
  graveyard at the time of starting a purge.
- Filtering out directory entries which are added while we're sending
  the initial state of the directory in the Watcher API.
- Debugging.

The sequence value is initialized to 0, and set to the current journal
offset at the time of committing a transaction (in
Mutations::apply_mutations). During merging, the common case of EmitLeft
or Keep preserves sequences; for the more interesting cases where a new
item is emitted, we use the minimum of the two items' sequences, since
the sequence number is most useful when it represents the "birth" of
some object. This only comes into play for extents and allocations.

Tested: Unit.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532301
Original-Revision: 7fdbb267bedbd88267505085c836476092940575
GitOrigin-RevId: 856875c7ed66f2fe608051346c8fb0cf09d4cb0d
Change-Id: Ie1c3c26e01b9158a66dbc336b55095fb008b84ea