[roll] Roll fuchsia [cm][routing] Make directory rights tests generic
Defines a CommonRightsTest providing tests of routing
of directory capabilities with restricted rights and
moves a subset of the existing routing-with-rights
unit tests into it. Runs these tests for both
ComponentManager and for the static routing analyzer.

The remaining tests will be run for the static analyzer
once it has support for framework capabilities.

Original-Bug: 61861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523861
Original-Revision: 65869d8244bea58f7a2962931222023dac0d7f9d
GitOrigin-RevId: e745cebda93ca6d9b811b922491ad60e35daae5d
Change-Id: I3849155a143e606462128274d477ca0532449d0e