[roll] Roll fuchsia [mold] Fixed rare partial updates bug.
The recursive unique element counting alrogithm implemented to take
advantage of the sorted array was incorrect, but this was hard to
notice since it worked in most cases and tended to err on the
conservative side.

This CL introduces a fix for it and, as a bonus, replaces a last
cover emptiness check.

Testing: fx test mold-tests && fx test surpass-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537061
Original-Revision: 4cd8e891e8538430481d089fe16bd2510a4e7582
GitOrigin-RevId: e2afdba962e7d003554822e55147dac54a43337a
Change-Id: I2f860337fa82fec304e2c0a47ab33fb623c72ffc