[roll] Roll fuchsia [fidl][llcpp] Remove unnecessary templating on client side
`fidl::ControlBlock` only needs to know the fact that the object under
management is a `ClientBase`.

`fidl::Client` needs to know the precise `ClientImpl` type in order to
expose proper `ClientImpl* operator->()` operators, but otherwise do not
need to rely on the `ClientImpl` type for its internal logic.

This makes it easier to insert the unified `ClientImpl`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/520841
Original-Revision: 13d0b61092531d15bd0709e788768c3efd45c307
GitOrigin-RevId: 5dcb3255891059b48d1f413907cd89bb73dfbaa9
Change-Id: I22c22cbde8dfbc9afedc7f3ce7cd9a782cd72c5f