[roll] Roll fuchsia [input][shortcut] Support press-release shortcuts
It turned out that shortcuts that involve a press and a release of
a single (modifier) key didn't work because of a bug.

This change makes it possible to use such shortcuts, and adds
new tests to guard the functionality.

While I was there, reorganized a few things:
- changed BUILD.gn related rules to be next to each other,
  and test fixture names to be a little less similar.
- added `Latch` and pulled handled assertions outside of assert hooks,
  thus allowing assertions to fire multiple times with different
  "handled" states.  This allows us to observe the end state
  of the test instead each individual state.
- Renamed the shortcut manager package used in the hermetic tests.

Original-Fixed: 76252
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/526517
Original-Revision: 5c57cbf4ee7eed3e65f34bba2b347feb89b7fc4e
GitOrigin-RevId: 4ff47de9a96cf32b590bc2f4c54c417c704a44b0
Change-Id: Ifac675d8d9d58207e5ba5559b5b228a355560c1b