[roll] Roll fuchsia [starnix] Pass PipeTest.Seek
This test requires lseek to return ESPIPE for pipes. We still do not
implement the actual seek logic for regular files, but this CL prepares
us for doing that once we have a test that requires it.

Also, update the test filter to run more tests that we already pass.

Original-Bug: 76861
Original-Bug: 77376
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537761
Original-Revision: 3a45b108041381a803793370aecc87a418449dc8
GitOrigin-RevId: b5d521efe40ff7dee0958b6f5c5bcf23997b6916
Change-Id: I3652fae3f4368ab49f442f0dabed6f411a0fe747