[roll] Roll fuchsia [fidlconv] Cache derived resourceness during compilation
When converting method request/response types, fidlconv needs to
know whether the top level type is a resource in order to annotate
it correctly in the new method syntax.

Previously this was done in a fragile way by attempting to reconstruct
fidlc's resourceness logic.
This CL updates fidlc to store derived resourceness for request/response
structs, which fidlconv can then use to perform lookup on during
conversion instead of trying to recompute the resourceness.

This approach breaks abstraction somewhat by exposing Library internals
to fidlconv, but this is considered OK given it is only temporary.

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler
--gtest_filter='ConverterTests.AliasOfResource'

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536938
Original-Revision: 421c68be18939169368563f043111510650e37fe
GitOrigin-RevId: 263ea2f7b4bfc4ad7d9ddce003387d5efdb5e3a9
Change-Id: I7b50d6c03d7d86304a7dad88a6df6db934b073e5