[roll] Roll fuchsia [fidl][hlcpp] C++14 constexpr flexible enum members
This changes the way we represent flexible enums in the HLCPP bindings
so that enum members can be used in switch statements. For a FIDL enum:
  flexible enum Foo : uint32 {
    A = 1;
    B = 2;
    C = 3;
  };

we would like to be able to do:
  switch (foo) {
    case Foo::A:
      // Something
      break;
    case Foo::B:
      // Something
      break;
    default:
      // Something else
  }

but this wasn't possible with our previous bindings. Flexible union
members were static const but not constexpr. That's because C++14 didn't
let us declare constexpr the member values out of line and they couldn't
be declared inline because they were declared before the class
representing the flexible union was complete.

This CL introduces a type fidl::internal::FlexibleEnumValue<Enum, Type>
that holds a value of the Type and can be implicitly converted to Enum.
This way we can have inline constexpr definitions of the enum members as
for example:
  static constexpr FlexibleEnumValue<Foo, uint32_t> A{1};

When we drop C++14 support we should return to the previous model as
it's a little simpler and clearer, even though this change doesn't
introduce any runtime overhead.

Original-Bug: 77240
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/535166
Original-Revision: 15adda7f734fee438851bfc81aad7203854f1f92
GitOrigin-RevId: e0e6e75e03e65944a0906014a9be8f200188dead
Change-Id: I0c2b432bced555b346bf2738b606bb1998a2d317