[roll] Roll fuchsia [system-updater] Reland Migrate system-updater to a v2 component.
This change migrates the system-updater component from the software
delivery stack into the core realm as part of the v1-v2 migration
project. System-updater is an eagerly started component that provides
access to the fuchsia.update.installer.Installer service.

As part of this migration the component depends on several services
provided by other components in the SWD stack. These are routed through
core.cml to the component. In addition to this this component relies on
a set of features & directories:
- build-info
- deprecated-misc-storage
- isolated-persistence-storage
- pkgfs/system

Relanding this change moves the system-updater routing in core into a
shard so that it does not have to be included in recovery images. These
don't require the updater and are space limited and cannot support the
eager requirement of the component.

Original-Bug: 73992
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534309
Original-Revision: 4f3d9c891bf5946d85bed80ec72d6f26b4d4d94d
GitOrigin-RevId: 90a3a46ce9724b72ee9d8c55eb73e4e6d28d012c
Change-Id: Idc0636f005ccc4e004864371e2584be7ed45fd6a