[roll] Roll infra/recipes [build] Strip most of api.build.test() helper
It was far too general-purpose and had a wide variety of
responsibilities that should really be handled by separate helper
functions. It was also the only easy way to mock the output of
testsharder, so it was often used even in places where it really
shouldn't have been used, e.g. many recipes did things like

  yield api.checkout.test("foo") + api.build.test("foo")

which is really awkward because the two functions are both intended to
be top-level test declarations and are not intended to be used together.
It's only incidental that the recipe engine even allows this.

So separate the testsharder mocking logic into a separate
`api.build.create_shards()` helper function that tests use explicitly
when they want to create shards.

`api.build.test()` now only contains logic related to mocking the
Buildbucket input, which I'll move into a `buildbucket_util` helper
function in a follow-up since it's not specific to the build recipe
module. At that point we can delete `api.build.test()` entirely.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/538282
Original-Revision: 4c15479857a19c5fc78508e2746a4c65e5c336da
GitOrigin-RevId: 934678e181e09381c7f2596f8257d2ba0c808aa2
Change-Id: I07ed63be833d847fee84e15c910c33f5c770d0e0