[roll] Roll fuchsia [pkg-resolver] Cache TUF metadata with non-auto repos
Previously, we would try to update TUF every time before we try to
resolve a package. This can generate a significant amount of load during
an OTA, since we would update for each package. This patch fixes this by
allowing the TUF metadata to be cached for 5 minutes, which should
significantly reduce the load on the server.

Original-Fixed: 77442
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/538013
Original-Revision: c8a2a422f7684f810af88217fa74ff2d374fdd64
GitOrigin-RevId: 85462c67b1e023fa8941bbf8eb876e2848c995f2
Change-Id: I8e83d131d293b8f8c54e59397d5689f79ac19ebc