[roll] Roll fuchsia [scenic][ui-input] f.u.pointer: Combine win with first message in exclusive mode
Makes it so that when injecting with exclusive mode the initial message
delivered to the fup client includes the "win" message for the
interaction. Before this the win would be sent separately, usually after
another hanging-get interaction.

The corresponding fix for the hit tested mode will come in a follow-up
CL.

Original-Bug: 76233
Test: fx test input_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536911
Original-Revision: 3d03bca1b9327aa9572d50aad9be19729ceb4fb2
GitOrigin-RevId: 67fc00155961da21ecf80c595ea5fb290d061228
Change-Id: I99b3b7144c78fa06feb5a469a1f45a03cef30de2