[roll] Roll fuchsia [forensics] Add internal FIDL protocol for migration data
Feedback's components are being combined into a single monolith that
will serve all the Feedback protocols. The data in namespaces of the
original components will need to be migrated into the namespace of the
new component in order to preserve persistent state.

This protocol will be internal to Feedback (and not in the SDK) and
short lived (it will be deleted once the merge is complete).

Original-Bug: 71830

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536908
Original-Revision: 6d3970234f7342db00cd7184bf01d1da90736b48
GitOrigin-RevId: 6346bc47257ad98ea62067536b89924bb9b5f12a
Change-Id: Iaab8db79f21b34f3468c4c9d0fdfb02a0aebc5db