[roll] Roll fuchsia [zbitl] ignore/take View error before exiting scope
This change updates several uses of zbitl::View in the tree in which
we forget to `ignore_error()` before exiting for a non-iteration
failure reason. In not doing this, we risk crashing a program in which
we could have otherwise more gracefully reported an error.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537305
Original-Revision: 92dd48f3c593f5b61dc18ee7f08cb18e9fb6c2e9
GitOrigin-RevId: 182dfc08f2776a03f1e1905a71eb7f2d6012a4f4
Change-Id: I116584345265204b01b1fee16229733de68e23af