[roll] Roll fuchsia [bt][at] Change the way arguments are parsed.
This CL fixes two issues.

Formerly, arguments were eagerly converted from strings to ints if
possible while parsing them.  This meant that if a string were expected
but the encountered argument were "1", the parser would convert this to
an int and the later attempt to match the AST to the command definition
would fail.  This CL removes this eager conversion and stores all
arguments as strings until attempt to convert them to typed structs
using the command definitions.

This CL also introduces the ability for lists to have optional elements.
Formerly an argument list like "0,,1" could only be interpreted as
strings, since the middle element is not actually a valid int.  Now
definitions can allow lists to contain optional elements and an empty
string will be  interpreted as None.  As part of that change the ability
for argument lists to have a trailing comma has been removed.  This is
wasn't actually part of the spec, and makes it impossible to tell if,
for example "AT+CMD=1," is one argument or two.

Original-Bug: 73374, 74304
Testing: fx test src/connectivity/lib/at-commands
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537108
Original-Revision: c894779f76ce06bf120504d7e2e2e4eccba49f04
GitOrigin-RevId: 91ba3dc7cd470062dd633105b46a3c2fff59a3c4
Change-Id: I33d9ef3cb4c7bac4c33fdcac5c2e119289fc68e0