[roll] Roll fuchsia [policy] Add * to VmexResource allowlist
This matches our policy for the
`deprecated-ambient-replace-as-executable` job policy.

We also need to remove the test that verifies that we deny VmexResource
to components -- on eng builds, it's impossible to create such a
component that wouldn't be allowed to access VmexResource.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/535981
Original-Revision: 2a19cee649014c6b12775435dd056a93dffb65b9
GitOrigin-RevId: 5a3b48e7a0ca58da0c9bb57bed39fc82bf9b616d
Change-Id: I54865c6de65108d1b1091ea20dadb9919a39f2ca