[roll] Roll fuchsia [brcmfmac] Fix dropped scan results race condition
When starting a scan we associate a sync ID with each scan, this is sent
to firmware and each scan result contains this sync ID to ensure that
the scan result belongs to the currently ongoing scan. The sync ID is
not stored until the scan has started, this means that it's vulnerable
to a race condition where if the thread that starts the scanning gets
suspended before it stores the sync ID results could arrive and they
would be matched against an incorrect sync ID. This would cause the
results to be dropped.

This fixes the issue by introducing a lock such that the whole process
of starting a scan completes before we start processing results.

Test: Manual testing able to reproduce without fix, not with fix
Original-Bug: fxbug.dev/77630
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536032
Original-Revision: 55bcf7ce1ef84c3b267ee128566294f08325700a
GitOrigin-RevId: edc01a24e8ed6e41dc0170562a930263dbad3bb5
Change-Id: I03779fa38822994be0171f186da3c604ac4f6341