[roll] Roll fuchsia [botanist] Run subprocess without piping stdin
Botanist uses the subprocess golib in tools/lib/subprocess to run
subprocesses. This library installs a handler that SIGKILLs all child
processes when a SIGTERM is received; it accomplishes this by sending a
SIGKILL to a PGID. Unfortunately, https://fxrev.dev/498141 breaks this
behavior because it does not set a PGID when stdin is being rerouted
from os.Stdin. This led to cases where botanist would get a SIGTERM and
stop the serial server, but testrunner would continue to run. This led
to the attached flake. This CL gets around this issue by removing
os.Stdin piping to testrunner, which isn't actually necessary.

Original-Bug: 72696

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/538989
Original-Revision: 23fe2ac8cd51ffbfe06d90e7e813a511ae4deeed
GitOrigin-RevId: feed7c81200db8d01dc70cb484e7d379014e3a33
Change-Id: I6a253e40de3d72e62688ac50b1cf1e58e097bd29