[roll] Roll fuchsia [fidlc][lsc] Make handle rights flag a no-op
This CL enables the "enable_handle_rights" flag always, by making the
actual flag a no-op. This means the remaining usages of the flag can
be deleted.

Handle rights works in all bindings now and with the exception of Dart
GIDL tests is functionally complete. Handle rights is no longer an
experimental feature and it seems like the appropriate time to quietly
enable the functionality in anticipation of a more visible launch.

fx lsc presubmit was run on this CL

Original-Bug: 77846
Original-Bug: 41920
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537461
Original-Revision: 86d9e1a071e8161ed3494fa6f7f18a6405868aad
GitOrigin-RevId: b190cc9efb882b8e8dbf7cd8b322cbcb28fb6416
Change-Id: I3df05f553df460b33870af30e581c04aed37e97e