[roll] Roll fuchsia [bt][hfp] Update DmtfCode values and make CallState flexible.
DmtfCode values now map directly to the ASCII values of each variant.
This can be used by an implementation to cast directly to an ASCII
character.

CallState is now flexible to allow it to be expanded in the future.

Includes minor documentation improvements.

This CL was created in response to feedback on fxrev.dev/538241

Original-Bug: 64498
Test: API definition changes only.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539029
Original-Revision: 17f5cf6e11611e464dd7f78fc4309e5a7d1442d5
GitOrigin-RevId: 09276e73a1e432561d9698ca3521f69d2730ccfd
Change-Id: I165d6572c077e2e2f7bb2deaf9a05c0d5a1a3a29