[roll] Roll fuchsia [debugger] Implicitly find variables on const this.
Previously the code did not strip C-V qualifiers when computing the
collection type for the "this" pointer. This meant that attempting to
print variables on "this" without qualification for const functions
failed. The situation is particularly noticeable for printing captured
variables in lambdas, which are const by default.

Original-Bug: 77910
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539021
Original-Revision: 2e65ea753170ce51a4df752bf8a0356feca5e9df
GitOrigin-RevId: 7e8db57e9d04db30ca738a49a4f39b2f89c17c96
Change-Id: If1f5351abd0f312191d43d28769c1217afaafbc1