[roll] Roll fuchsia [debugger] Highlight disabled breakpoints.
When one has a lot of breakpoints with some disabled, it can be
difficult to see which ones are disabled in the table. This highlights
disabled breakpoints by making the "false" string red so the table is
easier to scan.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539281
Original-Revision: c4ca1c02c6cdbc60f7b0eab352cdefe4b6ea2bad
GitOrigin-RevId: 6a697762e5056bb05ec3b4ba22b3475deb7c8481
Change-Id: I8615c2f2fa1e86aa05aa09a7921152df413c83d6