[roll] Roll fuchsia [blobfs][test] Deflake BlobfsFragmentationTest
The test was checking for specific metric values from Cobalt,
but blobfs does the metrics flushing to Cobalt in a background thread,
and it is not guaranteed that all of the metrics are updated atomically,
so partial results were possible.

Changed the test to wait until the metrics were the expected value, on a
deadline, which should help reduce flake.

Original-Bug: 75799
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/522521
Original-Revision: 2bdde615d3b61151d0f0ace505ad7031914b8536
GitOrigin-RevId: bc30e2b2fc9e50366d7c11831c5ff6eda24d0457
Change-Id: Ic4229abf710906ff8b64dba46798874c79ba323e