[roll] Roll fuchsia [ot-stack] Move lib-ot-core from public_deps to deps
This allows us to reuse the apps target to add a different lib-ot-core
flavor. This is needed so that we can create a different ot-stack
which depends on lib-ot-core-legacy-extensions.

Original-Bug: 73057
Tests: Build goes through, and functionally it's a no-op.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539441
Original-Revision: 2a6c62562b78c1f8eb7f7cd6d402811db727cb70
GitOrigin-RevId: f4268bb7fb371639065df97b6f920e59f336802d
Change-Id: Icae97a5e0b9f34655eae18c550f07b897e30ad16