[roll] Roll infra/recipes [archive] Add archive module to handle isolate uploads and downloads.
This moves isolate uploading to this new module which will eventually
handle uploading to CAS instead, and it also converts all isolate uploads
and downloads into just using the two functions in this module.
This will make it easy to switch from isolate to cas because there will
only be one place that needs to be switched. The only other place the
isolated recipe_module is used is in recipes/fuchsia/fuchsia.py where it
accesses the isolated.server.

Original-Bug: 62939
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/538802
Original-Revision: 37f52d990229920e7c22490647b6ba7dfef046e7
GitOrigin-RevId: 5c0fc1ca3c38bcd2cf40d50dd425c1578af3d677
Change-Id: I79c3395842f8484c9e33fa609644e87f84d0ee85