[roll] Roll fuchsia [build][go] Add sdk_name option.
The go_binary template forwards variables to go_build,
which forwards it to sdk_host_tool. go_binary supports
sdk_category but not sdk_name.

We would like to export the fidlgen_llcpp (go_binary) tool to
the SDK under a different name, but keep the original name
in-tree. The existing templates support an output_name field,
but that would also change the in-tree name of the tool.

Original-Bug: 74665
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537112
Original-Revision: 4aef047fedd9b349cce44cfd7db3cebfead60a72
GitOrigin-RevId: 80a8c8da0ade3333253b89833d929fa411f83d41
Change-Id: Idbbf9a6d98f4955c214ce3a6c868f01314f27e54