[roll] Roll fuchsia [build] Remove non-chromebook-x64 vboot image
This seems to have been introduced at a time in which we had
"chromebook-x64" developers working off the SDK, which was only based
off of a core.x64 build: hence the need to smuggle in a chromebook-x64
image through the latter. That is no longer the case today, as the SDK
includes chromebook-x64 images directly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539063
Original-Revision: 103ffb35f9afe53947a146c25e929e49dc42753f
GitOrigin-RevId: e3d6d75c703b87d8acc964b3e9230a7a7faca521
Change-Id: I6c45eedbce72a8ba8d669fc0b6d663336e73883f