[roll] Roll fuchsia [bt][hci] add AdvertisingHandleMap
When communicating with the controller, we do so in terms of an
AdvertisingHandle. This means that we frequently need to convert between
a DeviceAddress and an AdvertisingHandle. AdvertisingHandleMap provides
a bidirectional mapping between a DeviceAddress and an
AdvertisingHandle, ensuring a 1:1 mapping between DeviceAddress and
AdvertisingHandle.

Original-Bug: 78081
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539146
Original-Revision: 7bbd424fc49cc046a021ba17745c9b5d05ab029b
GitOrigin-RevId: be388ca078e0b3aba7395f979bbfe20eb0455e69
Change-Id: I7967190390bba97eb61cf1450176d3f7cdc0ebde