[roll] Roll fuchsia [lowpan] Remove use of config file for lowpan
Temporary change, this is conflicting in name with internal config file.
Need to figure out the reason for conflict and possible resolution.

Functionally a no-op -- default values used when config is not passed
are same as the values mentioned in the config file.

Original-Bug: 78178
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539680
Original-Revision: 576782d49626aac4d7f1aefa610f740c7d9c3f84
GitOrigin-RevId: 8c5c440a8402318e986d9342f9b50e67b048f0c2
Change-Id: I5c79ae973a8572b342bd239f36f0fb04031ddaeb