[roll] Roll fuchsia [phys][page-table] Add support for specifying cache attrs on ARM64.
In particular, we allow regions to be mapped in as either cached or as
device memory. We additionally add support for setting up the ARM64 MAIR
register.

The infrastructure supports more types of cache attributes, but we don't
have a need for these yet.

The implementation requires that users of the library set up the MAIR
register to a library-defined value. This constraint will be lifted in
future if necessary.

Original-Bug: 67632
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533403
Original-Revision: 2143e1cfc834499c7ec22b8186604cabdd46b90f
GitOrigin-RevId: 202a0bd3bbd8e03e6638699ffa904ed66a4d8f13
Change-Id: Iccaa16282d479201aa0510ce44c2c6bfeba9383e