[roll] Roll fuchsia [lowpan-spinel-driver] Added cleanup mechanism to API lock
Previously, if an operation like a join or form were interrupted,
then there was no way to put the NCP back in a known state (because
the future controlling that process had been dropped). This change
creates a new type for the API lock that allows us to register
NCP requests that are to be sent when the future is dropped, allowing
us to return the NCP to a known state.

Test: fx test lowpan-spinel-driver-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536161
Original-Revision: b997c9131563e501fc5305e74ecf3dedd64134bb
GitOrigin-RevId: d529e09a08d660b8af832576a56ea942d24ca430
Change-Id: I23f03bac894af63f3075d6892936a43ebf1145d8