[roll] Roll fuchsia [test][debug] Signal when a debug VMO may be read
This is the first of two changes to add explicit signals to DebugData
processing.
This change updates the DebugData protocol and makes the minimal
changes needed to successfully build. This adds an empty protocol to
DebugData.Publish, and updates clients to leak the client end of the
channel. The channel is implicitly dropped when the client process
terminates, which signals DebugData.
Note that this does not update test manager's DebugData implementation
yet. The server side will be updated in the next change.

Original-Bug: 75686
Testing: CI dryrun, existing tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532986
Original-Revision: 9093da3a41b0c694cc6c1d92af070074a1d65dd0
GitOrigin-RevId: d882d0e5dcf68c44b444d9dd0665cdd06f5d953a
Change-Id: Iad6598c63debb2c37ea1e979ccf35c58aeb17b5f