[roll] Roll cobalt Attach contribution_id to observations
The contribution_id is generated, if required, when the observation is
added to the store regardless of the state of encryption. The id is then
assigned to the EncryptedMessage after the observation is pulled from
the store and encrypted (if necessary).

For private contributions that are broken up into multiple observations,
only the first observation receives a contribution id.

fxb/72254

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/538890
Original-Revision: 9066956a7539132e07c5e1105f0da0611f4daa0e
GitOrigin-RevId: fc43c591f39a8e5c208cc0f2588be34fb2616fac
Change-Id: Idb6ab9c98ae276d604deb101f77b95275bd6501e