[roll] Roll fuchsia [e1000] Use driver protocol methods rather than manual scanning
The e1000 driver scans the BAR registers in the function's config
space to figure out where its IO bar is. This isn't disallowed, but
it's not recommended and may be locked down in the future. The method
now uses PciGetBar to obtain the BARs and check if they are the IO
BAR.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540264
Original-Revision: 5571c93865ae3ded9b05d7a28dfe0e4a09057ebd
GitOrigin-RevId: 58011aa53e13f495cf419e321eb855d50d1f92ff
Change-Id: Ibdee06034d2c86c73d6d9d0159dd5c21c2557ec1