[roll] Roll fuchsia [flatland] Replace Scale with ImageSize
Remove the ability to set scale on a flatland transform
and instead directly set the size of an image-content
which is local to that image and must have integer
width and height values.

Testing: Existing tests that are modified to accommodate
the API change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534821
Original-Revision: eb239ce09a03479ef68ac10b87dfadf3761151da
GitOrigin-RevId: accfd2b36779629249caec5b621e2755497d4cf6
Change-Id: I1a260f2377df58db0152ee0c2c114ff49b692c7d