[roll] Roll fuchsia [system-update-checker] Make current channel immutable.
This CL modifies the behaviour of system-update-checker to be closer to
that of omaha-client. It removes use of current_channel.json and instead
determines current_channel once at startup based on either
channel-in-vbmeta or rewrite rules.

Current channel is now constant for each boot of the system on all
products.

Original-Bug: 59881
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/511718
Original-Revision: 5355b2999f99d9423a2a0c7e3df932d802dccb39
GitOrigin-RevId: cd6d68c48e971d21a1b00150baffbb47553814f7
Change-Id: I3207881ef15edcddb95574dbfde3e89eb556b83a