[roll] Roll fuchsia [graphis][vkcopy] Validate only last copy operation in benchmark
Mapping and validating the copy operation may take a nontrivial amount
of time, so move it out of the measured portion of the test.

Also add a flush before the copy and an invalidate after, since we don't
request a host coherent memory type. We should be doing a pipeline
barrier to VK_ACCESS_HOST_READ_BIT before the invalidate, but that can
happen in a later patch.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/523103
Original-Revision: 119d8ea9f85121b8ba8548b93eac9e0cbf90fbf4
GitOrigin-RevId: 99988f0b7a01753dd1ae6ade914d2a2ce4e7a6b7
Change-Id: I30eb03ba3ecef80f68a96862e7a1c8a3753fdd7e