[roll] Roll fuchsia [wlan][sim-device] Add proper Unbind()/Release semantics
In the Fuchsia driver model, a call to device_async_remove() causes the
unbind hook to be called on a device, and then recursively its children,
before causing the release hooks to be invoked.  We implement this
behavior here in FakeDevMgr.

Original-Bug: 76745
Test: fx test sim_device_test brcmfmac_sim_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539678
Original-Revision: a3dabddaa05979e5dcfc0884d4634dbbf4454cde
GitOrigin-RevId: e4d8ceb603854985e477a64798cd1540fd620812
Change-Id: I09d6b45b18cd0f40d5bed9e00bd8afe41285f36f