[roll] Roll fuchsia [lockdep] Fix out of bounds reads.
This CL changes the way the LockDependencySet Iterator works.
Previously, the operator++ would check if the iteration index was valid
before increasing, but after increasing it it could become invalid, for
example, if it was equal to kMaxLockDependencies - 1.

This would cause an out-of-bounds read inside the set.list_ array, which
was detected by KASAN.

The new version, increases the index before checking against
kMaxLockDependencies. The behavior is identical for all cases, except
for when index == kMaxLockDependencies. But that's ok, because you
shouldn't advance an iterator that is at the end.

Original-Bug: 78333
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540584
Original-Revision: 87780435bc8e7891cb5ba9597d6290df15dafcb7
GitOrigin-RevId: 270df1ed35b3693b7d6aa0604c9abb034e7217fc
Change-Id: Ibac498ed26a024e3c7421feeac0fd7b6657d5a09