[roll] Roll fuchsia [brcmfmac] Prevent spurious log about bus state
When the driver is initialized it will set the bus state to down.
Sometimes the variable used to keep track of this state would already be
initialized to that state when it was allocated. This resulted in an
error message about already being in that state. Since there is no
meaningful previous state it's functionally equivalent to just set the
variable right away without calling the helper function to do it.

Original-Bug: 78276
Test: Build, boot and verify connectivity, only affects initialization
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539876
Original-Revision: 1b2d3ee644a8cfd238c3ffc9ffc23c814fea7e7f
GitOrigin-RevId: d75cdb6bfaa0f12a8d24f6cb4f31b85d6f69c82a
Change-Id: Ia17f2e2e92e7a39850e1d63e187e76ec14db4c4f