[roll] Roll fuchsia [assembly] Truncate package files before writing them
In the case where a new package is smaller than the previous package, we
were not truncating the length of the file to match the new size,
leaving excess bytes at the end. We now truncate the file to a size of 0
before writing to them.

x = old package
y = new package

[ xxxxxxxxxxxxxxxxxxxxxxx ]
[ yyyyyyyyyyyyyyyyyyxxxxx ]

This was causing the generated files to be invalid packages, and was
easy caught by switch products (via fx set) in the same out directory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540884
Original-Revision: 7d26e46a8beba80959aa8a4be852761b9dc5c3cc
GitOrigin-RevId: 497668f39081d4d4e463d3e90c86971857641930
Change-Id: Ia022194b82592b3e8e7a129e75e036fa97fd5643