[roll] Roll fuchsia [memalloc] Add algorithm for finding normalized RAM ranges
In early physboot bootstrapping, we need scratch memory for the
bookkeeping of memory itself. Today we accomplish this by allocating
a sizable, guestimated chunk of the .bss for this purpose. This change,
however, introduces a slick means of efficiently inspecting the span of
arbitrarily-overlapping memory ranges provided in the ZBI to find all of
the encoded, maximally contiguous ranges of RAM within, from which we
will easily be able to find our scratch.

Original-Bug: 77359
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/538782
Original-Revision: 084035819d9bc55c5742d8cee506ac597540edc7
GitOrigin-RevId: a92f4c50717793b892326c9e1e99e75b56fd8f66
Change-Id: I36fd735278e6a4281aaeb847c325e1f9c3506341