[roll] Roll fuchsia [debugger] Start decoding some return values.
Add some infrastructure (currently unused outside of tests) to find
what the return value for a function call was. This will be hooked up in
a later pass to display results during certain stepping operations.

This adds a new ABI class on the ArchInfo structure. Previously the
ArchInfo was only set when there was a connection and we knew the
architecture. Now since this is used in more places, an "unknown"
default architecture is set to avoid lots of null checks. The bulk of
this change is setting up this context properly.

Currently supported are basic types on x64 returned in a single
register, and complex structs returned by value. ARM support lags and
will be filled in more in a later pass.

Original-Bug: 5720
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539361
Original-Revision: 33eab06068677be2dde7a7a591ff21e251523e0b
GitOrigin-RevId: 9591be3bc2348241239bca190cceb8fb22c3d4c9
Change-Id: I9c50249342d401b179c1f312efa824f06183fb68