[roll] Roll fuchsia [botanist] Implement device flash support
Teach botanist how to use flash.sh to flash a target device instead of
paving it. Note that this change does not actually enable the usage of
flashing, as that is gated on foundation flipping the bit to enable
fastboot on hardware. Additionally, this solution by itself is
insufficient, as flash.sh does not support in memory booting of bringup
or ZBI test images. Still, this is a good starting point.

Original-Bug: 78321
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540481
Original-Revision: eb1f8bd99d883c09faa05a57db5ebb4575a5c23b
GitOrigin-RevId: 13fcb12a60a30d06a482f1146c9c9dabf26cfc4e
Change-Id: I5dde02775f918dac7e5af6a8de9f9403f3b6753c