[roll] Roll fuchsia [debugger] Show return values after "finish"
Shows the return values (when decodable) when a "finish" command is
executed.

In addition to the current limitations of GetReturnValue, this currently
has some issues with ordering. If there are inline frames after the
physical frame being finished, the thread will continue after the
physical frame "finish" and the registers will be invalid. This
asynchronous problem also means the return value is printed after the
code listing for the next command (which seems off).

To remedy this in a future pass, the callback will need the ability to
suspend the step operations until it has completed.

Original-Bug: 5720
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540886
Original-Revision: fdd59605b510541631b1263d3ac946e3dbebd466
GitOrigin-RevId: 810346562fcdba62090ccf84919d5a327c5a88af
Change-Id: Ib437ee557ae9d0157547617dce28f017d4116b4e