[roll] Roll fuchsia [tools] Add fint to build_infra bundle
The infrastructure uses a version of fint built using the Go toolchain,
rather than Ninja, so we technically don't need to include fint here.
But we still want the build graph to be considered to be affected by
fint changes, because fint is always used by the infrastructure to build
fuchsia and it may affect later infrastructure steps that might be
skipped if the build graph was considered unaffected.

For example, ci.chromium.org/b/8844872126628243968 ran on a fint change
and skipped running tests. However, we'd still like all builders to run
tests on fint changes to make sure that fint built all the artifacts
required to run tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541103
Original-Revision: 3232fb8dda94e839023a4215f859d8b277ad5ebd
GitOrigin-RevId: d1d4242fe1d5b89a2404a52a0c0ee3543ca2a057
Change-Id: I8761644aee21907fb53b101c5a7a33e592c1d52b