[roll] Roll infra/recipes [recipe_bootstrap_test] Enable property versioning for child
Property versioning is now enabled on most builders, so we should test
that recipe_bootstrap works with property versioning enabled.

We now set the "recipe" property in the child builder's checked-in
properties file that's read when property versioing is enabled, so we no
longer need to set it here.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/541342
Original-Revision: fd5acfa1434210f8a3c5f4b485e5b48961f37441
GitOrigin-RevId: 95b9ff7e8a410cf32d2572646f4f6acba0588123
Change-Id: I11fa00885e38a9b1cf09adaa35079ae80bf484d9