[roll] Roll fuchsia [fint] Add fvdl_intree_test to list of never affected tests
This is an e2e test that depends on the fvm so it's technically
affected (at least according to the build graph) by a large fraction of
fuchsia.git changes despite the test not actually being sensitive to
most such changes. So it's not useful to automatically multiply it on
those changes, and considering affected also makes it more likely that a
build will exceed the maximum of 3 affected tests allowed before no
affected tests will be multiplied in a build.

Original-Bug: 68434
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541181
Original-Revision: 10b8681fab71ab84e179de0a50e75a64eeb11ad1
GitOrigin-RevId: 6e23cc44222ce4c451980eeb18a98cc2f9156d9a
Change-Id: I79fa796b61e53950b178e2e1c50d7b6e7f3286d1