[roll] Roll fuchsia [zxio][tests] Refactor file/directory server types
This updates the file and directory FIDL server implementations for
testing to use a common test base built on top of the LLCPP-generated
test base types. This also updates the base classes to emit explicit
failures on unexpected messages with the name of the message in the
failure message.

Original-Bug: 43267
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541102
Original-Revision: a4252bd5da81954f228379d158147b80807a91f3
GitOrigin-RevId: 7e4f96fda61510c566a98bbcbc7e8cd5562eb90f
Change-Id: I9c01fb4aebdf743c5c1026961cfa60cb8a909ba7