[roll] Roll fuchsia [rbe] prototype compile wrapper for rustc on RBE
build/rbe/rustc-remote-wrapper.sh

The intent is for this script to be used as a command prefix
in rustc tool invocations, using the tool-agnostic
fuchsia-rbe-action.sh.
This script scans the rust compile command for a variety of inputs
and outputs (including depfile).

Input sources are taken from a locally generated depfile.
We use -Zbinary-dep-depinfo to get the transitive set of dependencies
needed to be uploaded.

Uses ldd to find related shared libraries for the rustc binary.

Rust standard libraries are listed in dep-info.

--inputs,--output_files paths are normalized relative to $project_root
which is $FUCHSIA_DIR, in practice.

Missing: dependent clang toolchain binaries

Checked with shellcheck.

Tested: (manual invocation)
  [from output dir, command from Ninja build rule for libsyn.rlib]
  ../../build/rbe/rustc-remote-wrapper.sh -- \
  ../../prebuilt/third_party/rust/linux-x64/bin/rustc \
  --color=always \
  --crate-name syn \
  ../../third_party/rust_crates/vendor/syn/src/lib.rs \
  --crate-type rlib \
  --emit=dep-info=obj/third_party/rust_crates/libsyn-59df571bc0ccf4fd.rlib.d,link \
  ...

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/537821
Original-Revision: 979991fd404dc46ebe6a7efd1546cbc18c43bbb6
GitOrigin-RevId: bc2c2e94d77123b5dc6fd1ca7fb9d5a48c3d3ea8
Change-Id: I39e5146e99dbd50713fe38de84bc0f6ae3116af0