[roll] Roll fuchsia [debugger] Allow porting of stdio
Unify the APIs of BufferedZxSocket and BufferedFD which were almost the
same but differed in some details (FD started watching on Init, ZX
didn't). Remove the Init phase which made construction more complicated
and just assert that the handles passed to the constructor are valid.
The common API from these classes is separated into a new BufferedStream
class.

Add a typedef in debug_agent for the the platform type to use for stdio
streaming. Use this new BufferedStdioHandle instead of BufferedZxSocket.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540302
Original-Revision: c7685f324494d530c7ded78244546c4bb8e8dada
GitOrigin-RevId: ce20959a47dddade466ade2be1b4742f942ca6dd
Change-Id: I0eb63c817e7f7fffeb0a2674e3594a01c156414e