[roll] Roll fuchsia [fidl][llcpp] Introduce WireSharedClient
Start of implementation of use-after-free prevention in client bindings,
following the proposed steps in fxbug.dev/75485.

This CL copies the existing fidl::Client interface into
fidl::WireSharedClient, since we would be restricting functionality in
fidl::Client later on. Subsequent CLs would evolve the API of
WireSharedClient to where we want it to be.

Tests related to cloning are moved to use the fidl::WireSharedClient
type, since cloning is going to be a feature exclusive to that client
type.

Original-Bug: 75485
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/536788
Original-Revision: a05d59d02c6d655b83f68c054efd183e3a69a510
GitOrigin-RevId: 418691f5e6548fd1d548ef49523b2e24092d972e
Change-Id: I77a2dcab9a3c3eacc7dad5991f942d989af9e39e