[roll] Roll fuchsia [fidlfmt] Implement resource formatting in the new fidlfmt
A small parser bug is also fixed, wherein the attribute type on the
resource_definition property subtype was being ingested as a
AttributeListOld, rather than as a AttributeList variant.

This implementation is bit awkward, but that is okay, since there is
only a single usage of `resource_definition` in non-test FIDL code, a
state of affairs which is unlikely to change in the near future.  A more
permanent solution to this problem would be to add a
raw::ResourceProperties node as the property list containing child of
raw::ResourceDeclaration, but this is deemed undesirable for a language
feature that is likely to change substantially in the future anyway.
A reference bug for this improvement has been added at fxbug.dev/78477.

Test: fidldev test fidlc
Original-Bug: 73507
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541124
Original-Revision: 3d789e0ab119b9c25dafc8012a84e23248cc5ae2
GitOrigin-RevId: 28c0ab910452483b2f9c6a0f6aa67449b5ef3d1b
Change-Id: Id738026b213a92b7b93b7731d234c00d7737c520