[roll] Roll fuchsia [build] Use an empty script_executable string
The script_executable variable in the .gn [1] file tells GN how to
construct calls to scripts from exec_script() and action() rules. The
typical use case is that these scripts require an interpreter to run
(such as Python) and the script_executable variable is used to specify
the path to the interpreter. In Fuchsia's build we use a mix of scripts
that specify their interpret directly and programs that run natively. An
empty script_executable tells GN to just emit the path directly. Scripts
which need to modify the environment can specify 'env' directly
themselves.

This may mitigate
https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=61784 by reducing
the number of absolute paths we use in script invocations.

Original-Bug: 61784

1: https://gn.googlesource.com/gn/+/refs/heads/main/docs/reference.md#dotfile

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541621
Original-Revision: 89120bb9c25378c5801de1d75101f47346076d46
GitOrigin-RevId: 8bb6eed83f91c7466f22b4854f64f16974f18ea0
Change-Id: Ia96329483cb76dca55bd62393ded742a5dc765fe