[roll] Roll infra/recipes [tricium] Exclude submodules from analysis
Currently, updating a submodule version causes the tricium_analyze
recipe module to try to operate on that submodule's directory as if it
were a file, causing failures like
ci.chromium.org/b/8844877592711820176.

Eventually we might want to include changes to submodule files in the
analysis, but submodules are currently on used by Cobalt AFAIK, so we
can punt on that and just skip analyzing submodules.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/541663
Original-Revision: 3c6b1f05aaada60ee6dafd3fff94616cb2cb528d
GitOrigin-RevId: 8c5baf06447a0e34b36102b8361b4ce9b47e421a
Change-Id: Ic23972b4a980878cca1db5c2224514648e13e7dd