[roll] Roll fuchsia [wlan][policy] Remove expected state machine error
This change modifies the exit reason when we exceed the number of
connection attempts in the state machine. This is an expected condition,
so we shouldn't treat it as an error.
Returning an `Err()` here causes an error message to be logged.

This boils down to a logging-only change, because the contents of the
`ExitReason` is only used for logging, not flow control:
https://cs.opensource.google/fuchsia/fuchsia/+/master:src/connectivity/wlan/wlancfg/src/client/state_machine.rs;l=151-158;drc=80903f7b7ec4ed2a9b21b17c434f2ce48dae403d

Original-Bug: 78486
Test: fx test wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541862
Original-Revision: 37c9af724ce30fb6810bec7c0173674e9be158db
GitOrigin-RevId: 35d449929e76c576bd67713336659a1536c3383a
Change-Id: I21136e6827596eb97aac80483d5789651c90777a