[roll] Roll fuchsia [fxfs] Purge old layer files after compaction
This CL also fixes a deadlock that could be triggered when zeroing the
journal.  It was taking the same lock that maybe_extend_journal_file was
taking, so I've moved things around so that extending and zeroing the
journal file is in the same place.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541163
Original-Revision: ed80766f92129e413ba5de3a774280ee5091c08b
GitOrigin-RevId: 13bf4257f0a0ee78a5d19637ec4969065602544c
Change-Id: I09335a6d7e877ede69f261cf4d5c793a4452bcf7