[roll] Roll fuchsia [fxfs] Wrap the filesystem with something that will close
Wrap the filesystem in something such that when it is dropped, it
automatically calls close on the filesystem if it hasn't already been
called.

This will go some way to stopping crashes in tests where close isn't
called.

To fix this properly though, we need to flip the strong/weak references
around so that the strong relationship is from child to parent and then
we also need to add strong references to any object stores that have
data that needs flushing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541301
Original-Revision: 6821cf0b89f76e6dbf64058af60488fb9b0debb3
GitOrigin-RevId: d06879d573529f2dc49e5ef9c969e30452d4e218
Change-Id: Ia6b7ae31deaa4a834b67d0c6c1802cce77d0463c