[roll] Roll fuchsia [sys] Make *ServiceDecl.source singular.
Prior to this change, the native versions of OfferServiceDecl and
ExposeServiceDecl had a Vec of sources, unlike
{Offer,Expose}ProtocolDecl and friends. In theory, this paved the way
for a FIDL service to be constructed from multiple sources, but that
functionality was never implemented and isn't currently planned.

Since the FIDL versions of OfferServiceDecl and ExposeServiceDecl only
had single source, it was not possible to do a 1:1 conversion between
the native and FIDL versions of these types, which is exactly what I
need to do as part of Dynamic Offers.

We can revisit this functionality in the future, but if we do, the
type conversion layer is probably not where the service aggregation
belongs.

Original-Bug: fxbug.dev/71881

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542083
Original-Revision: 00a95faaa2b413aea6da22757ac09cc911159317
GitOrigin-RevId: 34a0848ef2438b7b993ec0671b8a834c6cc6f726
Change-Id: Ie3e91209fdf55b7b4744c270a293bda69030365e