[roll] Roll fuchsia [gidl][rust] Add test to detect regressions in Rust GIDL build with no tests
Currently, there are often build issues with GIDL in various bindings
when tests are disabled. This can make it hard to disable tests for
debugging. This CL adds a test that always builds rust GIDL output with
an empty GIDL input file.

The idea is that this would be done for other languages as well. I'm
starting with rust to test out the idea so that work is not wasted doing
other bindings the wrong way if there are issues with this approach.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/532001
Original-Revision: 9708585968fd738e459ed7047abeb263a625acbf
GitOrigin-RevId: a3d13c59dba144408e950e9be25fb2863213e661
Change-Id: I807a2d2dad509ebe211a1da4e58d23657763abd8