[roll] Roll fuchsia [fidl][dart] Reduce handle rights when there are extra rights
When more than the required rights are passed to a FIDL decoder, the
decoder is expected to reduce the rights. This CL implements this logic
in Dart.

The test is written without GIDL because sufficient capability to
implement this in GIDL doesn't exist yet in dart (zx_object_get_info).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542263
Original-Revision: d3d2893bb1747f04bbf5a717258e52c2d6f61e6c
GitOrigin-RevId: ba70556501c826b7ea3e9f0dfdfea3aa3a0e76fa
Change-Id: I02506d4422fb3f9786013c0c6962f69f131d2574