[roll] Roll fuchsia [flatland] Errors close channel after returning OnPresentProcessed
This CL changes the way errors are issued in flatland:
- Errors are now returned as a param of OnPresentProcessed instead of
Present callback. This means one less wakeup for the client to check
the results.
- All errors are now unrecoverable and close the channel. Since
illegal operations are client's fault and may result in updates that
are not clear to the client or Scenic, we decided to treat them all
as fatal.

Original-Fixed: 56869, 36166
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540601
Original-Revision: 7dee958eeb5df5f27b4c66bd8d8edcda9cd9bc77
GitOrigin-RevId: 4429e14b5dfe307603f0039530eb8db31181b7f0
Change-Id: I46b4746e20071cf75a4f13529f97f82dee25d36f