[roll] Roll fuchsia [goldfish-sync] Fix a heap use-after-free error.
In sync device SyncTimeline::CreateFence() function, if a
fence is deleted from an inactive SyncTimeline (so the timeline
is not in the SyncDevice anymore), deletion of the fence may
cause the SyncTimeline object to be deleted as well, which could
cause an heap use-after-free error when releasing the lock
in the callback.

This change fixes this by using a RefPtr to the SyncTimeline
in the callback instead.

TEST=added a minimal test case SyncDeviceTest.TimelineDestroyedAfterFenceClosed
     to goldfish-sync-device-test to cover this case.
Original-Fixed: 73217

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542381
Original-Revision: ce7fff87c023340ab5ddc306d344b1574a38f6c3
GitOrigin-RevId: dea5193b5c071ff40c393abf91fc08c0ca201f1e
Change-Id: I7e28891ddf05a135186a49da91774cd745cc738b