[roll] Roll fuchsia [artifactory] Use relative paths to the checkout for host tests.
As described in the attached bug, some host tests want to have
dependencies on prebuilts which are in the checkout root. This was
causing artifactory to try to append a path starting with "../.." to the
upload namespace (because the host test paths are relative to the build
dir). Setting the paths relative to the checkout root should ensure that
all the host tests and their deps get uploaded under the expected
namespace.

Original-Bug: 77782
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540262
Original-Revision: 3f0c02a48de02a35c0964f563c6844283b32f6e7
GitOrigin-RevId: 04be3b8d0313c6dbfa7b1b572c4be1673d1941a3
Change-Id: If6eb5afbf3df24bdbb813a8374415446b22cdcf7