[roll] Roll fuchsia [audio] Add turn_on_delay to driver APIs
We add turn_on_delay to both RingBuffer and Codec protocol.
The RingBuffer to provide it to clients of the StreamConfig or
Dai protocols (for instance audio_core or bluetooth) and Codec
to extract the information from hardware audio codecs and provide
it to glue drivers that may use it to inform their clients.

Test: aml-g12-tdm-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/535242
Original-Revision: ff0204f34aac496114dee842c723548abd9a15d1
GitOrigin-RevId: ac5b6667212f444095967ba2349d8bdd9f2bb956
Change-Id: I27ec8863b44421ee71de90865c593b902c938cae