[roll] Roll fuchsia [fidlfmt] Fix new formatter crash
Previously, the formatter made the incorrect assumption that, in the
case where a protocol has both `compose` and method clauses, the first
instance of the former would always precede the first instance of the
latter.  This is not necessarily true, so we now check to see which of
the two comes first.

Test: fidldev test fidlc
Original-Bug: 73507
Original-Fixed: 78688
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542823
Original-Revision: 7ba01d3d2a51fa1b726763fb15ce30f87ad9f7a4
GitOrigin-RevId: c0285dba82ee23481c455f719cd1462f96eb6f7b
Change-Id: I7855d0a9ee47c7be48c2c39e68ce6df030f3798b