[roll] Roll fuchsia [fvdl] Add retry when fvdl exited with unexpected fails.
If fvdl fails with exit code 1, it indicates a logic error, and the test
should fail. If fvdl exits with code > 1, it means either aemu crashed
or too slow to start or ssh connection is not up; when this happens, we'll retry running
fvdl.

Original-Bug: 78190
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541141
Original-Revision: f46b9dcbe8620231bf0a5c25ad51960cdf2fb1fb
GitOrigin-RevId: 17a297cec4d441d8e70b128a95af4978b2ea70e5
Change-Id: Id8704c543b8834b651b7a35a74c07280eaf7443d