[roll] Roll fuchsia [debug] add handler class for the debug automation
The automation handler is a class that sits between the executor and
the debugged thread. It makes it easier to test the automation
instruction executor, and later it will make testing the automation
process in the debugged thread easier as well.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541001
Original-Revision: 6e4f682c17ac205f9e0da817e998c4c122be6270
GitOrigin-RevId: f87da3e56bf5740cb02b4e0caa9913c91ce06974
Change-Id: I4b8d710a062022e4fb108ccd4acf9bcf65b934fa