[roll] Roll fuchsia [ui-input][ime] populate `Key` in KeyEventInjector handler
When the KeyEventInjector handler receives an injection
request that has a `KeyMeaning` but no `Key`, attempt to
populate the `Key` from the `KeyMeaning`.

This improves compatibility in the case where a virtual
keyboard implementation does not provide a `Key`, but
a web app wants to process `KeyboardEvent.Code`.

While there: remove unneeded `clone()` of a `KeyEvent`.

Original-Bug: 78402
Test: fx test keyboard3_integration_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542392
Original-Revision: b00ff298b024636d44560ac261befcb27ad4216f
GitOrigin-RevId: 5488e88bdbd74a46d900970669ec11d995a57707
Change-Id: I09333b59137f62c1e45e13aafc58f9b5cfdf38f4