[roll] Roll fuchsia [netdevice] Add migration driver stub
Add an initial stubbed out netdevice migration driver. The driver
implements EthernetIfcProtocol, so that it can be the child of an
ethernet device, and NetworkDeviceImplProtocol, so that it can be the
parent of the netdevice driver. This migration driver is intended to
stand in for the generic Ethernet driver such that concrete ethernet
devices, like WLAN, can have a netdevice descendent driver without the
concrete device driver having to be rewritten in terms of
fuchsia.hardware.network.device banjo.

Original-Bug: 42640
Tested: fx test netdevice-migration-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540402
Original-Revision: acec6ccb39d35d5d4c8b755f28e12f0d13323c28
GitOrigin-RevId: b7ff5c0093988345c816aa9a32d9910b5ea67db8
Change-Id: I6fceed357bdf0b6a20d205ecce1e93b82a8d38f0