[roll] Roll fuchsia [memalloc] Make MemRangeStream work with an arbitrary number of arrays
This will allows us to populate an allocator with an arbitrary number of
range arrays. It will be more convenient to pass in three arrays, as
opposed to the previous two: at that point, might as well make it `n`.

Original-Bug: 77359
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542302
Original-Revision: 969ee448352a2c879efc58ecd058b298d19cd00b
GitOrigin-RevId: db5b141f6661a33cc22bc399221f12d20e45db86
Change-Id: I8759a719e00a2edbc1a7883b951e87f00905058a