[roll] Roll fuchsia [pkg-cache] Unify dynamic/retained package indices
The pkg-cache needs to track 2 mutable indices of packages, both the
dynamic index and the retained package index. While these indices need
to be tracked independently, the pkg-cache often needs to operate
against one or the other (or both), and it would be inconvenient to pass
around references to both and decide which index to operate against at
each call site. So, this change implements and uses a new PackageIndex
type that composes a DynamicIndex and RetainedIndex, exposing a similar
API surface as the inner indices that delegates to the appropriate inner
index.

Original-Fixed: 77361
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539875
Original-Revision: bafdb9a093a6ed64a99e0994f990356f1323aeeb
GitOrigin-RevId: dd1f214a11ad11a71ad66d8f550fa01a99b1279d
Change-Id: I11fd7e13ffae54f4bfd769a9fc01f62097e2c9ab