[roll] Roll fuchsia [debugger] Use new GetConcreteType calls.
Use the new GetConcreteTypeAs<> and the GetConcreteType variant that
takes a LazySymbol to clean up some call sites.

Remove the GetConcreteType "convenience wrapper" from ExprValue. It was
not actually more convenient, it just wrapped the call with the opposite
order of evaluation and didn't support the new variants.

Change a couple of symbol "AsFoo" calls to the new "As<Foo>" but only
where I was already changing code. The rest of these will be converted
later.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/543141
Original-Revision: 04d69add89e55d68ca9fdd8f84381acec6734452
GitOrigin-RevId: 6019cc3a357ab682dbd9aa3cea19a7b7641a6627
Change-Id: Ibedbc31ed35ba040d5b18ac91c0058a4d245e37b