[roll] Roll fuchsia [reland][time] Remove the zx_clock_get syscall.
We had to rollback the original due to a prebuilt that was
using an SDK well outside the 6 week compatibility window
and so still contained a zx_clock_get reference
(ref b/189107475).

That problem is being resolved with the help of the DX team
so we should be good to reland after the long weekend.

This is a reland of 22c0ce0edeb4a1c328aad341cfca6d42e9661643

Original change's description:
> [time] Remove the zx_clock_get syscall.
>
> Following RFC-0008 we removed all the users of the kernel UTC clock in
> F2, so it should now be possible to finally remove the zx_clock_adjust
> and zx_clock_get syscalls. zx_clock_adjust was covered in fxr/515400.
>
> Original-Bug: 74471
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/516981

Original-Bug: 74471
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/535269
Original-Revision: 9d290f0a03b0e4aa687da3c7ebc7a12da811574a
GitOrigin-RevId: 67a47ad0ddd2b8e9d8aa2e3c76c7289f220ec5c7
Change-Id: I78ae8a25897f04a667ffdd008aa8ff947b4e229c