[roll] Roll fuchsia [x86][acpi] Publish all devices in the ACPI tables.
This CL introduces a new acpi::Manager class, which is responsible for
the three phases of device discovery, bus type inference, and
publishing.

For now, these devices have no bind properties associated with them.

Currently, the old code in acpi-nswalk.cc is left intact. We will remove
it as drivers move out of the board driver.

Additionally, some basic infrastructure for testing ACPI functionality is
added, including a mock of (some) of the functionality exposed in the
Acpi class. Currently we just have a few tests that verify that ACPI is
interacted with correctly and all the devices we expect to be published
are published.

Original-Bug: 78348
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/543309
Original-Revision: c7c9eea28f3fc8e6433b6dc159c05c229b286e16
GitOrigin-RevId: 1bb8b1e25e0e9241e227c6e384d8f091bb1f1819
Change-Id: I51568541b0122355bad7c9d211b9ad1176a79654