[roll] Roll fuchsia [fidl][llcpp] WireSharedClient: Rename Unbind to AsyncTeardown.
Unbind is understood as extracting the channel from the client and
providing it back to the user. This doesn't match the current semantics
in LLCPP, which starts the asynchronous procedure of shutting down the
client. "Teardown" is a more accurate name.

I considered naming it |AsyncClose|, but it's not as accurate as the
client owns the channel. For example, the channel will not be closed
if there are hanging synchronous calls, independent of the unbinding
process.

In the future, we may be able to re-introduce an asynchronus
`void AsyncUnbind(fit::callback<void(fidl::ClientEnd<Protocol>)>)`
that better matches user expectations.

Also updates the documentation of WireSharedClient to talk about the
two-phase shutdown pattern.

Original-Bug: 75485
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539462
Original-Revision: f9ef22f89ddd077e0c89a22dee64c861d8af4107
GitOrigin-RevId: 3174070e933b015c48cc48a876c6502a706bfacb
Change-Id: I956805bad082c315cf66e9d4e94b56067639bb12