[roll] Roll infra/recipes [autocorrelator] Only re-raise known exception types
Turns out this code has a bug:

  try:
    ...
  except StepFailure as exc:
    raise type(exc)("...")

If `type(exc)` is a subclass of `StepFailure`, it may have an overridden
`__init__` method with a different signature from that of
`StepFailure.__init__`. This causes failures such as
https://ci.chromium.org/b/8845382453036456928, where we see a confusing
`TypeError` message in the summary markdown rather than the error
message for the actual root cause.

So instead, we coerce all StepFailure subclasses to the nearest parent
class among StepFailure, StepWarning, and InfraFailure. This does mean
throwing away other error context encoded in the subclass, but since
this is pretty much always the final exception we raise and there
generally won't be any more error handling around it, the summary
markdown is all that matters anyway.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/544023
Original-Revision: c22fa9abfcae22eb712a968d7ee7875914d9f870
GitOrigin-RevId: 5931cba6516a025f8b55fb0be6a387d2f89e31b1
Change-Id: I0d344e6b1a71bfd08d5ff90500e5193e03dd0521