[roll] Roll fuchsia [bt][hci-emulator-client] Fix test build issue
Thanks to computerdruid@'s efforts to build ALL the tests with cargo,
he found that we were not including these (ignored, but building) tests
in the GN dependency tree. Due to this, we forgot to move one of the
constants out of fuchsia_bluetooth that was only used in these tests.
Fix both issues.

Original-Bug: 72932
Test: Build and run test, see that the test builds successfully.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/543795
Original-Revision: b62cf41c33acdd93d97738e17ec375ba490553fc
GitOrigin-RevId: 360b24637ab7dc2392c878fdec1425e4ebab6f5a
Change-Id: Icd99167839ac260e62394bf51cf04d9a7dcebc18