[roll] Roll fuchsia [fx][daemon] Add `ffx daemon log --follow`
This allows the ffx daemon log to be followed for changes. In addition,
this also tweaks how to print the last few lines of the log file in
order to not buffer the entire log file in memory, and makes sure that
both features work together.

Original-Bug: 75500
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541982
Original-Revision: da9d587bc74ba0150cef66cd290a5e096add007d
GitOrigin-RevId: e7f9203edc74d4a2b46c0ccc0b2bccf5f1b6d18b
Change-Id: Ie47dda7c2ed29b4189623fa64c137d6220ef0bf2