[roll] Roll fuchsia [bt][rfcomm] Implement RfcommTest FIDL API
Update RFCOMM Session logic to support receiving the test requests,
and update handlers to expect appropriate responses from the remote
peer.

For the Disconnect request, we expect either a UA or DM response.
In either case, the RFCOMM Session with the peer will be terminated.

For the RLS request, we can only send it for established channels. In
this case, the peer is expected to respond by echoing the command with
the same parameters.

Original-Bug: 61306
Test: fx test bt-rfcomm-tests
Test: PTS RFCOMM/DEVA-DEVB/RFC/BV-04-C, RFCOMM/DEVA-DEVB/RFC/BV-14-C

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540942
Original-Revision: 09d38112695952fc7f1c335b9c85be418318d1ad
GitOrigin-RevId: 9d349449b51b4c664862637055f78e2f88bc6720
Change-Id: Ibbcecff1024f7cd5480d10beccec34532f8b711b