[roll] Roll fuchsia [fvm] Rename version variables.
Brings the naming of the FVM variables and constants in line with the
current policy. The old policy had a "format version" and a "software
revision".

The format version is the same as the major version. This was renamed
which should have no effect.

The software revision has been changed to minor version. In the old
scheme the software revision was monotonically increasing, while minor
versions reset for major version changes. However, since we have never
changed the major version, this change will also have no practical
effect.

No changes to on-disk formats or any values stored in them have been
made. Only the internal names in the software and some metrics cales
have change to be more consistent, and the policy for future
modifications has changed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/543244
Original-Revision: 30eed39ced54b18491c1b43ca3bfdf559b6cf1db
GitOrigin-RevId: 1d07dcc90f31efdf6d84829238c722f081ca94cb
Change-Id: I47ecef37e7cac8566cb5d2a190ae7daa1a1cac71