[roll] Roll fuchsia [netemul] Use relative URLs
Wherever possible in netemul, replace absolute static URLs with relative
ones so that the netemul sandbox can be embedded in test packages. It's
currently not possible to use relative URLs with RealmBuilder, so that
will come in a future change.

Also, remove the netemul sandbox package and leave it as a plain
component so that it can be embedded in test packages.

Original-Bug: 78313
Test: fx test -d netstack-integration-tests-v2
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540722
Original-Revision: fd4dff28912c224845f491a032a9c6c7b23a8839
GitOrigin-RevId: 39dabbee588480e10206a50a9788dd2abd0f1431
Change-Id: I6ef908a2421e3864ffa23eadd8d8349daf32828c