[roll] Roll fuchsia [SetUI][FIDL] Use jobs for Factory Reset Interface.
This changelist replaces the existing fidl processor-based
implementation of the FIDL Factory Reset Interface with
a version based on the new job infrastructure.

This is a reland of https://fxrev.dev/523986 and includes an additional
fix to avoid a crash on bad input.

Original-Bug: 70534
Original-Fixed: 78846
Test: fx test -o setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544064
Original-Revision: 320f60f0db9af92298bc895498a2e1b0574a89c5
GitOrigin-RevId: 697be81e723f61c771c23a1e19a10fef65809968
Change-Id: Ic3b24650eae8ebd159b20323c53187834a937851