[roll] Roll fuchsia [zxio] Add support for channels in zxio_create()
This adds support to the zxio_create() entry point for wrapping a
channel speaking the fuchsia.io.Node protocol by calling the Describe
method and then processing the resulting description.

This also adds a zxio_create_with_on_open() variant for channels that
are expecting an incoming OnOpen event, most commonly as the result of
calling fuchsia.io.Directory/Open() on a parent with the
OPEN_FLAG_DESCRIBE flag specified.

The new entry points are not used in fdio or outside of tests in this
change. fdio still requires type information to allocate appropriate
storage for the zxio objects, so it is using the zxio/cpp/inception.h
defined entry points.

Original-Bug: 43267

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542998
Original-Revision: dd25ccc1accd9391e500cccec10178713cb81fff
GitOrigin-RevId: a42012650e41658ba7f343e722b4b2d7f9a16e2d
Change-Id: I828afc3e6fdf7d2ab6b6e24e17c00bbaa33ca090