[roll] Roll fuchsia [bt][hci]: remove local address parameter to OnIncomingConnection
fxr/533529 introduced a change to pass the local address in a new
connection to LowEnergyAdvertiser::OnIncomingConnection(...). Upon
review and discussion with the team, this is the wrong approach. We will
instead use the results of the `HCI_LE_Advertising_Set_Removed` command
and reconcile the `AdvertisingHandle` and `ConnectionHandle` there.

Original-Bug: 76213
Test: fx test //src/connectivity/bluetooth/core
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/543428
Original-Revision: a1ce7ad9cb39ef2d0e3f09cbdbe24dcc8bedebca
GitOrigin-RevId: 6c55f4df3c51e25c0c62ea50ccc93d8268642c2d
Change-Id: Ief77fd0e88627f75ee1041475640de5918297275