[roll] Roll fuchsia [flutter] do not merge cml files in build step
In preparation for starting work on the CFv2 version of the
flutter and dart runners we need to have flutter and dart
components to launch. The current build rules attempt to merge
the cmx file with a file provided by the runner. This breaks
when we are trying to build a v2 component so we will just skip
merging the files for cml files. In the future we will need to
either include a shard.cml file or do some other type of merge
in this step but for now we just leave the manifest as is.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545041
Original-Revision: 6db61e4eaba5ec542e4efe3dcecae58176aee2e9
GitOrigin-RevId: 365117b43219a4c7f4bc8b6a4caa8174e2b0c4d3
Change-Id: I71b6c53323e197b8e170dac9072d60a370a58da1