[roll] Roll fuchsia [kernel][hypervisor][arm64] Avoid reuse of hypercall numbers.
EL2 currently has two modes: a hypervisor enabled mode (where hypercalls
such as "guest resume" are available to the host), and a disabled mode
(where such hypercalls are not available).

In the current setup, hypercall 3 is allocated to the "on" hypercall
when the hypervisor is disabled, and "off" when it is enabled.

This can lead to problems if host code (due to its own bug) gets
confused as to the state of the hypervisor. In particular, the "on"
hypercall expects a stack pointer for EL2 to use, while the "off"
hypercall has no parameters. A kernel thread inadvertently calling "on"
when it expects to call "off" will start using whatever value happens to
be in the register `x1` as its stack.

We update the hypercall numbers to avoid reuse, and update the invalid
hypercalls to return ZX_ERR_BAD_STATE.

Original-Bug: 77932
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544302
Original-Revision: 135f3f1d0d84ffc93126e01b43024f88c792072e
GitOrigin-RevId: e629867ec502649ee4974d813e1987266e0d1b9c
Change-Id: I5a4dd567cf1a71b82c4e2b49abe7cef6a2e9bad4