[roll] Roll fuchsia [netstack][fdio] Add stream socket shutdown fidl method
Note that this commit does not change behavior, except that before, some
bsd socket functions implemented as a fidl call (eg, listen or connect)
could be executed out of order if it was followed by a shutdown. This is
not a problem anymore, now that shutdown is also a fidl call: order is
guaranteed.

Use an environment variable to have the tests exercising the old ABI
(where shutdown is done by fdio shutting down the zx_socket, instead of
using the new fidl method).

Original-Bug: 78129
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540782
Original-Revision: 38152709ff314fe48532cf66dfa68f6f1f104da9
GitOrigin-RevId: 33b5d8dbd20b23338cd058190da5f29a2b26066e
Change-Id: I8fe78a939e2345933ae081ec77681ace6eba5d06