[roll] Roll fuchsia [bt][testing] rename FakeController::RespondWithSuccess to FakeController::RespondWithCommandComplete
FakeController::RespondWithCommandComplete uses the actual name for the
event that is actually going back to the host. There is another
FakeController::RespondWithCommandComplete which takes a ByteBuffer as a
second parameter. This change adds an overloaded function which takes in
a hci::StatusCode allowing developers to send back success or any other
hci::StatusCode with the same method.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/543429
Original-Revision: 58421e78806f7e8fb29dfc7c59deba0d1bff19ee
GitOrigin-RevId: a08004e393740a3385b3ecc49a66ce4a86ba2650
Change-Id: Ia8a06cede241dd444a266e7e939ebcc99dc25be6