[roll] Roll fuchsia [rbe] Avoid listing unstable directories as inputs
reclient assumes that the contents of `--input` directories are stable.
Allow-listing files is always preferred and safer.
Attempt to snapshot the files that are intended: .a, .o, .so (these are
not temporary, and will not disappear randomly).

There can still be instability if multiple concurrent build actions
write outputs to the same directory.
Adding files to the same dir could reduce a cache-hit rate,
and removing (temporary) files could cause the reproxy process
to fail with "file not found."

It is safe to over-specify inputs.
If more precision is desired in the future, then -Lnative flags will need
to be examined.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544221
Original-Revision: 5b0213420a4e01705e0189c7312e73b0e38a6eb6
GitOrigin-RevId: 4561c44c3f69cfc73176366b1d76b48db62529d6
Change-Id: I53769c8054b961fbddcfca6253ac866ba52d0ef7