[roll] Roll fuchsia [brcmfmac] Fix tracking of disconnect origin
The driver currently tracks the origin and type of disconnects to
determine what kind of notification to send to SME when the disconnect
has completed. The variable to track was not properly reset in one place
leading the driver to sometimes think that any disconnect event from
firmware was actually the result of a disconnect initiated by SME. This
in turn resulted in the driver sending a confirm event rather than an
indication event to SME. SME then ignored this event because it made no
sense and thus SME was not aware that the device had been disconnected
from the network and so of course no attempts to reconnect were
initiated.

Fix this by clearing the tracking variable in the one place it was
missing. Also added a sim test to catch this issue.

Original-Bug: b/191086331
Test: brcmfmac_sim_test, wlan_tests, added sim test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544828
Original-Revision: 870008719ba5def3b857bcc584c0de4db5a10438
GitOrigin-RevId: 29811f5a38f37b10d13a21eb7d42634fd9c79f46
Change-Id: Ic0b10f8cb53826a6d46a52ee2441011f1272462d