[roll] Roll fuchsia [build] Specify path to asan libclang_rt relative to the root build dir
When building for asan, we pass a path to the appropriate clang runtime
support library as a rustflags argument. The path is computed by running
the clang binary which outputs an absolute path. This rebases the path
to be relative to the root build dir per our best practices:
https://fuchsia.dev/fuchsia-src/concepts/build_system/best_practices?hl=en#prefer-relative-paths-from-rebase-path

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544181
Original-Revision: f1370eaafd3a710457b236fc6d00f0e0b5bc34b1
GitOrigin-RevId: dc5d07881eb498b55acb4c68458ec050f2fa9fd0
Change-Id: If4080ce9c38d39c4925335ac363dfcc7960b7b04