[roll] Roll fuchsia [fxfs] Change the flush strategy and block-size
This changes the flush strategy for object stores so that we will
eventually do full compactions.

To get it it to work, I had to increase the block size for persistent
layers and object stores as otherwise I was finding that flushing ended
up using up so much journal space that it would trigger an immediate
flush again which led to an infinite loop.

I also add to address a TODO in the allocator so that now we'll sync the
journal and flush the device to make available pending deallocations.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541884
Original-Revision: 7c41524282b348f6d554f5eff61980cd51e4ddc9
GitOrigin-RevId: a0539fcd637d0cf82fdfd647154768cf110fb27d
Change-Id: I20c8288d6b7cae916e0e962af5f77229e523db85