[roll] Roll fuchsia [fxfs] Clean up on major compaction
During a major compaction (one which goes down to the base layer), we
have an opportunity to remove records that no longer need to be kept
around, such as tombstone records and extent deletions.

We need to do this as a second pass after the compacted layers are
merged, rather than during merging, since otherwise we might discard a
record that should have applied over another (e.g. an overlapping extent
deletion and allocation). This is implemented with a
MajorCompactionIterator that wraps the regular merging iterator and
discards records when possible.

This also includes a bug fix: When tombstoning files, we didn't actually
delete all of the extents. Fixed and added a test.

Tested: Unit test.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542881
Original-Revision: d5f9226ce261ec886119bcb44446b35a2a523e3b
GitOrigin-RevId: f180bc9a159806f416fe688cb3f5231526656d5e
Change-Id: Iab1b4ca3e4c50187d0e63275c5007ed21ceca869