[roll] Roll fuchsia [fidlc] Parse modifiers into new raw::Modifiers AST node in new syntax
There are two reasons for this change:
    * It enables new lints that check the order of modifiers.
    * It moves some decisions about the meaning of modifiers (for
    example, what is the default strictness of some layout?) to the
    compiler, where they seem to be a conceptually better fit than the
    parser.

Of note, this CL has to account for an edge case like:

    struct { foo flexible; };

This is done by modifying the ParseCompoundIdentifier() method to take
the first Identifier as an argument, so that we may convert a
maybe_modifier of "flexible" in this position into the identifier for
a NamedLayoutReference.  Because we have to "lookahead" to deduce what
the last token in a set of possible modifiers means, and could
possibly produce multiple raw AST nodes when doing modifier parsing, the
old ParseModifiers() method is removed, and modifier parsing is done
inside of the ParseTypeConstructorNew() method.

Test: fidldev test fidlc
Original-Bug: 72121
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544823
Original-Revision: b36991abac6bc4231d07740a434b46122181e288
GitOrigin-RevId: 6fb5142ef840117e1eabdbd1bd637110e759c586
Change-Id: Ibe8326ed027aad94f390ac73d03dcc99e831d8e6