[roll] Roll fuchsia [testrunner] Retry tests after timeouts
Previously there was some conflation of the shard timeout for a
multiplied test and the per-test timeout, so we would always stop
running a test after *either* of those timeouts occurred.

However, we only want to stop running the test if the
shard timeout elapses, but keep retrying (depending on the
`RunAlgorithm`) if one run of the test exceeds the per-test timeout.
Otherwise we don't retry a test after a timeout, even if the test is
allowed to run more times, e.g. in
https://ci.chromium.org/b/8844182070029860976.

So I disentangled the two types of timeouts, and renamed the multiplied
shard timeout from `TimeoutSecs` to `StopRepeatingAfterSecs` to make it
clear that it's not really a timeout, but a soft deadline where it
shouldn't be considered a failure if it's exceeded. I also refactored to
stop using context cancelation to enforce the shard time limit. It's
just a soft target duration and we care more about collecting a
reasonable amount of pass/fail data points than about not exceeding the
time limit, so it doesn't make too much sense to cancel a test midway
through when we hit that time limit. Instead we'll let the test
complete, and then stop rerunning it.

I also added a new `clock` library to make it easier to mock
`time.Now()` for the purpose of testing this new logic.

I tested by intentionally adding a hang to a test to cause it to
timeout, and running a build to make sure it gets retried:
https://ci.chromium.org/b/8844056435482049984

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544922
Original-Revision: 28ae1d4970a3b7806f70efe66a2cb555c3c7b6cb
GitOrigin-RevId: 5da9dcd8781886bdce3b43467542428c2b974b49
Change-Id: I96a1945f3e148d83186902c6cc5f77e1c441ea77