[roll] Roll fuchsia [blobfs] Rename version variables.
Brings the naming of the minfs variables and constants in line with the
current policy. The old policy had a "format version" and a "software
revision".

The format version is the same as the major version. This was renamed
which should have no effect.

The software revision has been changed to minor version. In the old
scheme the software revision was monotonically increasing, while minor
versions reset for major version changes. However, since we have never
changed the major version, this change will also have no practical
effect.

No changes to on-disk formats or any values stored in them have been
made. Only the internal names in the software and some metrics cales
have change to be more consistent, and the policy for future
modifications has changed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545685
Original-Revision: e7212fa30c43e521d4fbf50dbb622678b1b101e0
GitOrigin-RevId: 894fe9d64404779018f3225762cc60f210abbaad
Change-Id: I45f0c3425cd294b1e32486584b8ff8441c20ac1c