[roll] Roll experiences [Ermine][Quicksettings] Remove bluetooth from quicksettings
The fuchsia.bluetooth.control library is deprecated and is planned for removal at the end of the quarter. This CL removes the bluetooth quicksetting which depends on this library. Bluetooth will be reintroduced using fuchsia.bluetooth.sys in ermine next.

Original-Fixed: 73844

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/545886
Original-Revision: 5a103839bc58045d9ed98182eeea1031e69a2fdb
GitOrigin-RevId: 9460a9a1e30db539d0a3cfca4a2152f08ac90ef6
Change-Id: Idec4fb1a37044ca4df12fa8fe895ff100452d28d