[roll] Roll infra/recipes [symbolize] Set a timeout on symbolizer steps
We try to present stdio for debugging when the symbolizer fails, but if
a build hits its top-level timeout due to the symbolizer being slow or
hanging, the error handling code won't run and we won't be able to see
stdio.

So set a timeout on just the symbolizer step, so we'll be able to see
stdio if it hangs.

Original-Bug: 79052
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/545889
Original-Revision: 67b0550c5d8450ec0448f1f5dc8de6a753f3e087
GitOrigin-RevId: 4725ed419440795c458d379ff0dd35fb9ffbc75e
Change-Id: Ic576c4da0f89411fda228e57f76723e418f75025