[roll] Roll infra/recipes [symbolize] Set a timeout on symbolizer steps
We try to present stdio for debugging when the symbolizer fails, but if
a build hits its top-level timeout due to the symbolizer being slow or
hanging, the error handling code won't run and we won't be able to see
stdio.

So set a timeout on just the symbolizer step, so we'll be able to see
stdio if it hangs.

Original-Bug: 79052
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/545889
Original-Revision: 67b0550c5d8450ec0448f1f5dc8de6a753f3e087
GitOrigin-RevId: 4725ed419440795c458d379ff0dd35fb9ffbc75e
Change-Id: I8f0d03345e3147b28b1d73b79e942e45aee7599b