[roll] Roll fuchsia [starnix] Use single_run for starnix manager children
This change updates the starnix manager to use the newly introduced
`SingleRun` collection durability for the components it runs. This means
that the children are started when they are created (don't need to bind
to them to start them), and they are automatically deleted when they are
stopped (e.g., when they exit).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/540323
Original-Revision: 37c4ed744ff360a4b2e69b5607bc03ff4c8c1417
GitOrigin-RevId: 1872b5deab22457068d6958103a5188fd73ae8fa
Change-Id: I223fa2b885c8b201556d0208ba86d24aac6600c7