[roll] Roll fuchsia [zxio] Add a sync zxio_open() version that wraps the opened object
This adds a zxio_open() call which is similar to the existing
zxio_open_async() call but also synchronously processes the OnOpen event
and wraps the result in a zxio object. This allows users of the library
to traverse a filesystem tree using exclusively zxio primitives.

Original-Bug: 43267
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/543963
Original-Revision: c5984e9966bb873cdd7033d4128de399d1742ed8
GitOrigin-RevId: 8a4b96357602faf5046265e47480223b1c372e93
Change-Id: I7fa3a8123a5f6c2e44cf2b2efeab665cec3f561d