[roll] Roll fuchsia [build] Remove unused config-data declaration.
The config-data directory is currently defined as used but not routed to
system-metrics-logger.cm from core.cm. This means if system-metrics-logger
attempted to access /config/data in a given build it wouldn't be able to read
the directory.

Currently it appears that the code doesn't use /config/data so it
appears to be a no-op to remove the usage.

This is being fixed as part of an effort to enable static capability
flow analysis into the build.

Context: https://fuchsia-review.googlesource.com/c/fuchsia/+/545001
Original-Bug: 61862
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545954
Original-Revision: 4ed6eed710b41dd34520bceb38757c9fff643518
GitOrigin-RevId: 6d37cfef652a63534d10f6dd1a4d12d19d1235d9
Change-Id: I09c84274f7ae9b3137b5ac06e243d6a29872f50a