[roll] Roll fuchsia [zircon] s/ZX_INFO_PROCESS_V2/ZX_INFO_PROCESS in general cases
For general uses of the process info API, we drop the "_V2" suffix, as
they are both aliases of one another. We keep the mention of v2 only at
sites that need to be aware of both the v1 and v2 variants.

Original-Bug: 30751
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546261
Original-Revision: 35f08541deeebed9ec341f622a6ac324979b50c2
GitOrigin-RevId: be0800520ffb048f4ca81938c3cde2e3e51d4de4
Change-Id: Ib69e872738a09c583ec26942fcb6e53a1b33037f