[roll] Roll fuchsia [kernel][x86] Disable preemption around idle loop spin
Fuchsia's idle thread, when using HLT rather than MONITOR/MWAIT,
will execute a small halt-poll before going to sleep. While
polling, it uses |halt_interlock| as a signal to avoid wakeup
IPIs - while polling, the idle thread sets |halt_interlock|
to '1'. Before sending a wakeup IPI, a sender attempts to
compare-exchange it from '1' -> '0'; if it succeeds, it elides
a wakeup IPI. If it fails, an IPI is sent. Just before a core
goes to HLT, it compare-exchanges '1'->'2', re-enabling wakeup
IPIs.

Unfortunately, there was an oversight in the initial
implementation of this optimization - when Fuchsia switches away
from the idle thread on a CPU after setting |halt_interlock|
to '1' but before setting it to either '0' (seeing a wakeup)
or to '2' (been woken up), the next wakeup IPI will be skipped,
which was not intended.

Instead, disable preemption around spinning in the idle loop
and check for the preemption pending bit in each spinloop. If it
is set, break early and re-enable IPIs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/534521
Original-Revision: 430e8b413ab34ba553efbc17556f0fd528dce656
GitOrigin-RevId: 532dd13624e9040246bf220e893078ad4576cf1d
Change-Id: I0d3660eafc652813fda741afd2228fb5634acf32