[roll] Roll fuchsia [brcmfmac] Refactor and rename brcmf_cfg80211_change_iface()
Refactored and renamed brcmf_cfg80211_change_iface() to
brcmf_cfg80211_change_iface_mac_role() to reduce the ambiguity
between IF type and mac role in the syslog. IF type seems to
be associated with a model in the firmware whereas mac role is a FIDL
abstraction.

The previous code used a switch statement and flags to trigger if-else
branching. The refactor uses only a switch statement instead.

Test: None. Pure refactor and logging change only.
Original-Bug: 78905
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544769
Original-Revision: a1ed78b996bd791a70f050b43036bdc8e187711f
GitOrigin-RevId: cd572bcd54e2933be3c0220e6080353f2f021ba9
Change-Id: Id68db943b9aa757d9e06ba8fb0699d77d4fd4395