[roll] Roll fuchsia [ffx][config] Fix flakes when saving ffx config
Previously, multiple tests that manipulate ffx_config could racily
modify the underlying config file, which would cause the tests to flake
out. This patch fixes this issue by:

* Changing `Environment` to grab the global lock whenever the
environment file is saved or loaded.
* Changing `Environment` to atomically write to the environment file.

This change isn't the most efficient way to solve this problem, since we
only really need to lock the environment when we're reading and writing
to the same file, but we typically only encounter that in testing, so
the overhead of the global lock shouldn't impact non-testing use cases.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546181
Original-Revision: 5e29378be38ba45dbd5f66d0fb68282654970727
GitOrigin-RevId: 53ef6d885eb5d3572a893b0155abd08c21dd524f
Change-Id: I04158da8969def3cb619df4a5c77816703284940