[roll] Roll fuchsia [zircon][x86] map all possible segment selectors.
This CL modifies both the LDT and GDT setup code so that all possible
segment selectors are mapped.

The CL disables the LDT in x86_init_percpu by calling DisableLdt(),
which calls lldt with a null segment selector. All segment selectors
that refer to the LDT will fail with a general protection fault.

The CL also maps all the GDT entries. In zircon, the GDT limit is set to
the maximum possible value (size 0x10000), but only the first pages of
this are mapped, causing a segment selector to the end of the table to
cause a supervisor level page fault. The code now maps all the remaining
entries as copy-on-write, so future write accesses would cause a page
fault.

Before this CL, userspace could craft a segment selector that would
result in a supervisor page fault, causing the system to panic.

There are two set of tests added: a kernel unit test that tries to
access every page on the gdt, to make sure that it doesn't fault, and
userspace tests that try using selectors and make sure that the kernel
doesn't crash.

Original-Bug: 79060
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545006
Original-Revision: b0952827a18d87363ee7b5222bbdac810c58a0a4
GitOrigin-RevId: 01753d4b14409be7f7eb26f10e9064726f43d75d
Change-Id: I1da6ab2cdcf3426ce7f5767ebeade8cda0d5d121