[roll] Roll fuchsia [kernel][vm] Multi-generation LRU Stage 1
This is the first step in the implementation towards using a
multi-generational LRU queue for the page queues. The goal of this
change is to have a lock-free path for marking pages as accessed, but
without changing too much of the page queues API.

For not changing the page queues API this means that there is still a
fixed cadence to incrementing the MRU gen, and the idea of peeking the
pager backed queue is emulated, instead of having the LRU processing
directly perform eviction.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/515943
Original-Revision: 7be5a6eed7c06ad343b03a1783deb93b800b4888
GitOrigin-RevId: b390600ce72c68c0719c7ccd95c0b4ec72e491d4
Change-Id: I462f477a61affcbd804bdd1b02e3346032bbcc79