[roll] Roll fuchsia [ffx] initial support for socket based diagnostics
This the change to the diagnostics bridge FIDL protocol to add support
for sending non-truncated data over a socket. This is the minimal change
to add that support but keep the implementation/functionality of logging
the same today. In a subsequent change the plan is to make use of the
new input parameters as well as actually add support for sending data
over the socket.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/528681
Original-Revision: d40c749565ea821f786798ec8e24674ea69524ce
GitOrigin-RevId: 512b6a58a1528085cab9a3aabf84b65c77074bde
Change-Id: I8277614c93c139a7821172d6a02a7510dabb2641