[roll] Roll fuchsia [wlan][iwlwifi] Ensure alignment of SimMvm resp_buf_
SimMvm contains a member of type SimNvm, which has zero members and thus
is a 1-byte sized member.  The next member of SimMvm is resp_buf_, used
as a response data buffer and declared as a uint8_t array, which has an
alignment requirement of 1 byte (as its elements have 1-byte alignment);
thus the array's address starts at a 1 byte offset.

Later on, resp_buf_ is cast into other struct types, but as it has a 1
byte offset from a typical aligned value, this causes undefined behavior
(and UBSan errors) down the line when multi-byte struct members are read
or written in an unaligned manner from the aliased struct.

In this commit we move to allocating resp_buf_ dynamically, which
ensures that the buffer is aligned to at least the platform's strictest
fundamental alignment requirement.

Original-Bug: 46688
Test: fx test iwlwifi_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544601
Original-Revision: 34f209a4ddeee1185fa3e4039316a549d89eac53
GitOrigin-RevId: 898c0999285ee77d0cc04ed4af740c7a9d788274
Change-Id: I651b6429f93fc261c53525a1cb72be0ed2839111