[roll] Roll fuchsia [virtualization] Simplify PCI capability structures and set up.
PCI capabilities are currently passed in as an array of cap objects by
the subclass of PciDevice. When the PCI config space is read, the
current code proceeds by finding the correct cap struct, and then
emulating byte-by-byte reads on this struct, allowing it to patch cap
ids and next pointers as required.

This setup will cause a few difficulties in future CLs:

  * The PciDevice itself will want to advertise capabilities for MSI-X
    support directly, instead of relying on the subclass advertising
    it. This means passing in a static array of caps is not ideal.

  * The current byte-by-byte processing will cause difficulties
    when implementing write support for caps, which require atomic
    word writes (and not splitting up the write byte by byte).

This CL refactors the capability code to allow capabilities to be added
dynamically, and set up in more than one location. Additionally, we
update the code to perform cap next pointer patching when the cap is
installed, instead of during reads to the config space.

Original-Bug: 78022
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/538826
Original-Revision: cab0b3d9779875928aef9e324ce55ed84991b808
GitOrigin-RevId: 9d641d160bacbe028165520f9744e4f7305caf0e
Change-Id: I8d7a673c3c8116eb00d84cda1b239c1fde5402c9