[roll] Roll fuchsia [virtualization][vmm] Convert ReadBar/WriteBar methods to callbacks.
Currently when a PciBar is read or written to, the PciBar class
redirects the read/write to the ReadBar/WriteBar method on the parent
device. The parent device reads which BAR index was read/written, and
executes the appropriate action.

The current setup makes it difficult to allow different BARs to be
routed to different locations (e.g., one BAR for MSI-X interrupts to
the base PciDevice class, another to the derived class), and relies
on statically-allocated identifiers to determine which BAR was
read from / written to.

We update PciBar to instead take a callback for reads and writes,
allowing reads/writes to be directly routed to the appropriate handler.

Original-Bug: 78022
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/539542
Original-Revision: 3a39d59675b2766a831bd5afcc94c790d2a700e0
GitOrigin-RevId: f1ede6432ab0ebef847df03493676672acabaeea
Change-Id: I1db0ac83ebc3a2cb8193b3e8e5806f155e13b86f