[roll] Roll fuchsia [carnelian] Add flex factors to flex
This CL adds a mechanism for groups to store per-member information to
be used by a group arranger and uses that mechanism to implement flex
factors to be used in the Flex arranger.

In order to allow arrangers to change the size of groups and facets,
rename Facet::get_size() to Facet::calculate_size to better represent
the intent of the trait method. Also store the size of the of the facet
along with the location in FacetEntry. While making this changed, moved
FacetEntry into a more logical module.

Eventually calculate_size will allow for facets that scale to available
space, but no facets do that yet.

Introduce the GroupMemberData type alias to Box<dyn Any>. This allows
the arranger to store one bit of arbitrary data with each group member,
if needed for arrangement.

Create GroupEntry struct to track both a member of the group and any
GroupMemberData it might have.

Add an option to pass GroupMemberData to places where facets and groups
are added to groups.

Made the member_sizes slice mutable so that arrangers can change the
size of their members during size calculation.

Pass a matching slice of group member data to arrangers so they can
downcast them for needed per-member data.

Add flex factors to the flex arranger. While this has no effect on
facets, since none of them are yet able to adjust to a new size, it
works on other groups. See the one_third_two_thirds mode in
layout_gallery for an example. In in, two columns in one row are given
flex factors in order to produce a one-third, two-thirds layout. The
columns alignments are end on the left and start on the right to cause
the two test facets to abut.

Removed the caculate_group_size method on Scene, which wasn't used and
it's not sure it can really work outside of a full layout.

Updated all implementers of the Facet trait to match the new definition.

Added the OneThirdTwoThird mode to layout gallery and refactored the
implementation a bit.

Testing: manual plus unit tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544864
Original-Revision: ed70556a7654419b0fc9ac5fad29543f27ecc253
GitOrigin-RevId: e2c12a522d1813eac3453a2eafd16a9152f90979
Change-Id: I111123b903d4a72514694d7c2dcb1a9ad8c7f75a