[roll] Roll fuchsia [fs] Make easier to identify thread safety of Watchdog class.
Was already thread safe as implemented, this just rounds up as many
members as possible to be either protected by the main lock, marked
const, or moved to be method local variables instead. The remaining
exceptions are the sync_completion_t and the std::thread. Both are
types that are meant to support thread safety and have a narrrow usage
to provide easier auditing.

Original-Bug: 67659
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546280
Original-Revision: 793fdde7eb26bad3f620cec406ed5fd53da481f8
GitOrigin-RevId: da7753c5ef9c0b6ae7f92302bdbdb5b2e906b525
Change-Id: I64349add749f73c1d5679c4f1bda479d72620227