[roll] Roll fuchsia [scenic][ui-input] Only collect contenders up to the target
Currently for TOP_HIT_AND_ANCESTORS_IN_TARGET we collect contenders
all the way up to the context. This is incorrect. This CL changes it so
that we only collect contenders in the Target subgraph.

Also updates the DispatchPolicyTest to a proper unit test and to use
fuchsia.ui.pointer, so it can be used to catch this issue.

Original-Fixed: 79454
Original-Bug: 72919, 53634
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546550
Original-Revision: 19f85a24bc3c674c5695246788ba035eab60ae42
GitOrigin-RevId: 671f7bbb58f55bb4d793136ed4ad126215ebcfdf
Change-Id: I9ab6697721c88fe5cd30cc9414dc6b6b25b5d6bf