[roll] Roll fuchsia [diagnostics][triage] Unify multi-select and Option()
A "select:" entry with multiple selectors will now discard Missing
errors, the same as Option() does. This means that a selector
with a component moniker that's not found will no longer cause
a Missing error to be propagated, as long as a later selector
in the list finds a component.

Test: $ fx test //src/diagnostics
Original-Fixed: 79081
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545705
Original-Revision: e523be0651870a804ba2d20304032e4a3e6d207a
GitOrigin-RevId: fecdffd5d9d2b23cfc9865fb926c03cbfcedd881
Change-Id: Ic3c6c9ab6f2be908768f069d2a998e4a32eae336