[roll] Roll fuchsia [pci] Support making PCI devices composite with ACPI.
This CL adds a new list of PciBdf to PciPlatformInfo. For any device in
this list, we add a third composite fragment to the composite devices
which binds to the ACPI device that represents the device at that BDF.

For now, nothing populates this list, and so no devices are composite
with ACPI. A future CL will populate this list in the x86 board driver
and start creating composites.

Original-Bug: 78565
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/545150
Original-Revision: d97474833681a2385533ca1253c337e6c5407a2b
GitOrigin-RevId: 1364394d336c5de704ec6b91d185ce266dcc730d
Change-Id: I9c345446755fd082a379bc2068fe6af84ffd795b