[roll] Roll fuchsia [kernel][vm] Unify accessed harvesting
It is no longer neccessary to know anything about a page, in particular
whether it is in a particular kind of VMO or not, to update its
accessed time in the page queues. This change leverages that to remove
the accessed callback, directly calling MarkAccessed from the mmu code,
and then performing a single iteration over the arch aspace for both
page aging and page table reclamation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/533288
Original-Revision: b549b3a376856cc8870bb0aef815bb5b121d2d74
GitOrigin-RevId: cda9248f543d31c05c6a6027abc810e66beeea49
Change-Id: I758d3641d57c50adeb17a9c450f408000c30825f