[roll] Roll fuchsia [gn_sdk] Fix run_py.mako to work under Python3.
The template's #! was updated to refer to Python3 rather than Python2,
but the template content was not updated to be Python3 compatible,
seemingly because these tests are not run anywhere.

Fix the template's "print" syntax to be valid Python3.

Original-Bug: 60187
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546081
Original-Revision: f3ef79f99596dcd6ba5238fef1fb52f7dcff2c83
GitOrigin-RevId: 170d78c5ccf97e3d262a43584148489ceb2e1a5f
Change-Id: I8cefba2134c08619cd0b20408195b85bceb09bd3