[roll] Roll fuchsia [amlogic] Dump registers on RDMA stalls.
There is circumstantial evidence that the RDMA engine is stalling on
these devices, leading to "frozen" screens that get refreshed with old
content. In the event that we cannot make forward progress for 2
seconds, dump all the registers. This should give us a better idea as
to whether the device is being programmed incorrectly or the silicon
has a bug.

Original-Bug: 189400236
Test: build and e2e on local device.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/547132
Original-Revision: a1b1ec87cb781f338e00ccc98c460a6ac4fae93c
GitOrigin-RevId: d5f6e54fae93a0b063e3961296565b846d57f0ef
Change-Id: I3ff1acbe42b6f52adfbcb0e1b112bb7de3d6e779