[roll] Roll fuchsia [flatland] Use channel instead of eventpair for link tokens
We currently use eventpairs that can be duped which can mess up the
lifetime expectations. We explained this in the documentation, but
we should have a better of enforcing this. This CL enforces it by
using channel instead which cannot be duped. Events still work
through ObjectLinker to match with their endpoints.

Original-Fixed: 79509
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546908
Original-Revision: 8561e5b4cd2ba780b569adc0bebfb9136e87f9dd
GitOrigin-RevId: defa5ad985de1925d4f43dfebf10152f05ca55cd
Change-Id: Id20abc6483504899042ef3de05b8a06a1a040d25