[roll] Roll fuchsia [debugger] Eval DW_OP_piece for unspecified data.
DW_OP_piece can accept no data in which case the bytes of the output are
said to be undefined.

This writes 0's to the data. This allows these expressions to be
evaluated but will be misleading if the undefined data is displayed.
Tracking that data is issue 79547.

Fix a bug in DwarfExprEval to clear the result data when it is
reinitialized. This was triggered by the added test. It is not triggered
normally because we normally create a new object for every expression.

Add an upper bound on the size of a piece to prevent runaway memory usage
on corrupt data.

Original-Bug: 79528
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548221
Original-Revision: 3be465bb95d1b602675013471ec3cb3695bd8f62
GitOrigin-RevId: c87ece26455bfa8b8c82ba9269872aebcfeb3e55
Change-Id: I9c9ac5b94dcc067861bb81861dfb85366b7c9b0b