[roll] Roll fuchsia [usb] Reduce logging severity from aml_usb_phy_v2 driver
If the phy_rdy register is not set within a 1ms interval, the driver
currently logs to ERROR level, but otherwise continues normally. As this
does not appear to be a fatal condition, the log is reduced to WARNING
level. Other public drivers for this same hardware do not check the
phy_rdy register.

Original-Bug: 79564
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548351
Original-Revision: ec98d64d53afeb8bb3ae23f5621c2a40b262a3ea
GitOrigin-RevId: 56a0f4d6869ffb94870aac2b1e66cc83656eae37
Change-Id: Ib6b3307aa489e33d93a1e4552d38af959ab60a91