[roll] Roll fuchsia [modular] Present shell view directly without a ViewHost
Pass the view from InitializeSessionShell directly to the shell
ViewProvider, rather than creating a ViewHost and adding the shell
view as a child.

Ensures that the shell is focused when the sessionmgr view is
presented in basemgr.

Test: manual
Original-Fixed: 78739
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548143
Original-Revision: eb4e0d52264d115ca8a919a9fbbb67e59d30428a
GitOrigin-RevId: a981566d558d6886db91765c435700ef03f8c8ca
Change-Id: I7e5c6c88695f079b79c8a8993904b2a0fb4a866c