[roll] Roll fuchsia [fssh] Moving info logging to debug
Info level logging appears on stdout, which is intermingled with the
output of commands run on the device via ssh making the parsing of the
result difficult.

This moves the message to the debug level, so it can be used by
users to diagnose an issue, but not interfere with normal use.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548483
Original-Revision: 38b72f4dd7ffee1157e63d90ea3b0a635b716da5
GitOrigin-RevId: 26fd5e474e77b7e6eda6d90626fe2f7e60650bb6
Change-Id: If48656f73c9d37b1608aa2c2b7d68904912d8f93