[roll] Roll fuchsia [netstack] Record stats upon DHCP packet drops
The DHCP client silently discards frames which are deemed irrelevant.
This change adds logging and statistics to track each instance in
which a frame is so discarded. The motivation is to provide better
data for investigations when a device is unable to acquire an IP
address.

Original-Bug:77810

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/542021
Original-Revision: 0fbfac09db0bdf606f00d84d63825e9fdc3f9278
GitOrigin-RevId: 038206d9a1d2f468d02bcf3333f74117aae5524f
Change-Id: I435247a7434ab0bd0ec4730c37b201d943a0c617