[roll] Roll fuchsia [pkg-resolver] Improve blob fetch error reducer
When a package resolve encounters an error writing a content blob for a
package, pkg-cache may decide that the error was fatal and tear down the
Get() operation, which also closes all other concurrent blob writes for
that operation.  The pkg-resolver was simply returning the first blob
fetch error as the root cause of the resolve failure, but the
pkg-resolver may observe a cancelled fetch before the more useful error
that triggered the fetch to be cancelled, resulting in it sometimes
reporting a generic INTERNAL error for resolve requests that should have
failed with a more specific error.

This change modifies the package concurrent content blob fetch flow to
* wait for each blob fetch to succeed/fail with a cancelled error, or
* wait for any blob fetch to fail with a non-cancelled error
returning the more interesting error encountered, if any.

Original-Fixed: 79559

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548729
Original-Revision: 202e388289e5a44f471ccdd4eabd1bc66eefca25
GitOrigin-RevId: 1476000a1b0d204c947b34721a825a27dcf7f102
Change-Id: Ia752e75a1decd5b4454e7b9c4affc7a4f3e32e2a