[roll] Roll infra/recipes [checkout] Explicitly resolve manifest revision before checkout
Rather than letting `jiri import` resolve it, so that we have the
manifest revision available to emit as an output property for use by `fx
sync-to`.

This only results in a behavior change in the "local CI" code path (e.g.
a build triggered by a fuchsia.git commit). We have always done this
in tryjobs, and in integration CI we get the pre-resolved hash directly
from the build input.

I also renamed a few instances of "integration" to "manifest" in parts
of the code that are generic and capable of checking out any Jiri
workspace, rather than specific to the integration repository.

Original-Bug: 79014
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/548002
Original-Revision: e55d887703c840b36fb7faff224a27a6fa08d20e
GitOrigin-RevId: cbbeb4b4a314c2f88ff80587200cb8da837475d5
Change-Id: I970654622bc95c0aa13e935321e9cb94f758b257