[roll] Roll fuchsia [kernel][hypervisor][arm64] Don't reboot on unhandled exceptions.
The EL2 hypervisor attempts to reboot the system via a call to
`psci_system_reset`. Unfortunately, this call expects to be running at
EL1 and attempts to access the kernel heap via a kernel virtual address,
which isn't valid in EL2. This leads to another fault, and repeats
indefinitely.

Instead, change the action to just hang the processor, and rely on
Zircon's lockup detector and/or watchdog timer to eventually kick in.
Hanging also means the core's register state is preserved, which makes
it easier to track down the cause of the exception via DAP.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548969
Original-Revision: 98de5c48824720df1b8280dbcabdd91944724b7e
GitOrigin-RevId: 224c0d5a78463f4950d1d19e7073feb048170f3a
Change-Id: I2786db511bb48105cf423ccc629591552c9ea9b2