[roll] Roll fuchsia [starnix] Add basic futex support
This CL contains support for the basic WAIT and WAKE futex operations.
We put the thread to sleep in a port, which will let us interrupt the
wait to deliver signals in the future.

This CL does not check the value for WAIT yet. That check is difficult
without atomic reads from the child process, but we can probably get
something basic working in a followup CL.

Original-Bug: 74727
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548490
Original-Revision: 53dfe4ced1498210c638e9886ab965caab2bc105
GitOrigin-RevId: 7bfb95e2cf71167f785d18f0d38fcf5a9283a6a0
Change-Id: If286a06c702ee284578469049a2581557e59dbf2