[roll] Roll fuchsia [starnix] Fix CLONE_CHILD_CLEARTID
Previously, we would clear the child TID immediately when creating the
child via clone. However, this flag really means we should clear the
child TID when the child task is destroyed. It's the signal from the
kernel that the thread is gone.

Original-Bug: 77376
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548491
Original-Revision: 18a270a752c3904895c938dad380fad6417b8c9d
GitOrigin-RevId: d12dd91949a3a8c27035a2fe9e01a0f22867915c
Change-Id: Ie2ba4399b99abaf488711680f646d25dcb8265b3