[roll] Roll fuchsia [starnix] Add support for blocking I/O
This CL adds a mechanism by which read and write operations can block if
O_NONBLOCK is not set. They install an observer on the FsNode, which
gets notified when the node is readable (or writable). The threads sleep
in a port, which will let us interrupt the sleep to deliver signals in
the future.

Test: PipeTest.WriterSideCloses
Original-Bug: 77376
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548492
Original-Revision: 7df145ff61e58e3b9e00ef5918faff73c1c4f078
GitOrigin-RevId: f327310f51d7041cdd3a289031b70d2e75ba1c32
Change-Id: I83009eae697747b5a7c512a5014575cda4b90e12