[roll] Roll fuchsia [build] Remove -fexperimental-relative-c++-abi-vtables flags
Relative vtables is enabled by default on Fuchsia, so we can remove
these flags that explicitly enable it.

Note that we will still need to keep `use_cxx_relative_vtables` because
some external libraries like libneural_network_models.so are built with
some BUILD.gn files from the fuchsia source tree and may want to opt-out
of this feature for now. Once we confirm there are no external users
that disable this, we can then remove the GN flag and the config.

This also changes the `compiler_relative_vtables` config to either
explicitly enable/disable RV when included, depending on the value of
`use_cxx_relative_vtables`. This should not be used anywhere in the
source tree.

Original-Bug: 45314
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548141
Original-Revision: a94b8d119a5448c3c22225533239b1fb5f65af07
GitOrigin-RevId: 475b568c4276166e407e1500cc71fd210b1a4618
Change-Id: I252897cef13e635bb1d63c688c55ed8ea6b574f8