[roll] Roll fuchsia [audio_core] Rename ReadableStream::Buffer::gain_db
The current name is confusing: the term `gain_db` sounds like it should
represent gain that should be applied to the stream in the future. In fact,
it's the amount of gain that has *already* been applied to the stream.

Hence I renamed this field to `total_applied_gain_db`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544501
Original-Revision: c492b40dc86270e5e5b41951329ae41cfb7383b5
GitOrigin-RevId: 319d92ae8446186c91196b544f10162419475d22
Change-Id: I0f617158b9dee0e22eed2b77366ee21d7dffa00a