[roll] Roll fuchsia [wlan][iwlwifi] port mac_clear_assoc()
This is the last piece of code that enables the disassociation. The
changes include the Tx queue management and the STA state transition.

This is also the follow-up of fxr/56379. The mvm_sta ownership is
actually transferred into mvm->fw_id_to_mac_id[] while configure_bss.
So we cannot remove it while we are still using it.

Original-Bug: 36742, 49531
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412252
Original-Revision: 71fe1f79a9ee93dc237bf7b1e798abd4244a4237
GitOrigin-RevId: 203f79ca6186305b0fcea550787094369c17a89e
Change-Id: I7291f89305f27b68b9001c15df2f4b06bf12da0c