[roll] Roll fuchsia [fvm] Log FVM info on startup.
Enhances FVM logging on startup to include more information. It now
includes the full header information in addition to the free slices, as
well as all partition information.

The previous code attempted to log the partition information but instead
logged the pointer because of a coding bug.

The GetFvmLocked() was renamed GetHeaderLocked() to match GetHeader().
This tripped me up while writing this patch.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549903
Original-Revision: a6aee82d186b7e78cfd20860a328b487289625e3
GitOrigin-RevId: 6685f97e88f54e780ce220bcb2223e0cc5dc70bb
Change-Id: Iaba61b2cf6419b78e68ff66009e89cb4a2b31c08