[roll] Roll fuchsia [fidl][llcpp] Introduce WireClient
WireClient checks that it is always bound and destroyed on the
dispatcher thread. It is suitable for systems where the user owns the
binding object, which is a significant portion of non-driver LLCPP
usage, and all of HLCPP.

With both WireSharedClient and WireClient, we may deprecate the
fidl::Client which isn't speicalized to either use cases. Follow-up CLs
would add more documentation and migrate fidl::Client to either
fidl::WireClient or WireSharedClient depending on how it's used.

Original-Fixed: 75485

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544988
Original-Revision: 68eeea6f5d515f07a8bafcf295d8e9ff5975a69f
GitOrigin-RevId: b04b49f28b3408cee476dc84e281264e3634bfce
Change-Id: If81bdbaa777564ba6757809862ddc1c7a5f82541