[roll] Roll infra/recipes [checkout] Expose checkout details for use by `fx sync-to`
In order to reliably reproduce an infra checkout, `fx sync-to` basically
needs to follow the exact same steps as infra – i.e. check out the same
base revision of integration.git and then patch the same CL revision on
top. So we expose all the necessary information from the checkout recipe
module.

Also update fuchsia.py to pass on the "checkout_info" property from a
subbuild.

See fxbug.dev/68492#c5 and fxbug.dev/68492#c6 for a description of how
`fx sync-to` can use these properties to reproduce a build.

We'll no longer use the Jiri snapshot at all because it's not possible
to reproduce an integration.git tryjob's checkout from a Jiri snapshot,
as there's no way to determine the base integration revision on top of
which the change was patched. The output properties approach will
support this use case.

Original-Bug: 79014
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/548003
Original-Revision: bdad637168e756446a4e9d8705265cd08bf59dba
GitOrigin-RevId: 83e56acb306aa49c8cddde3d25648d7208da1ec8
Change-Id: I0b5848025be01d2b80ae3efce147d8b1603c7051