[roll] Roll fuchsia [virtualization] Add libvulkan.so.1 to romfs
Normally on Linux libvulkan.so.1 is the real name (e.g. soname) of the
vulkan loader, while libvulkan.so can be a symlink to the latest
version. We can use a variable from Vulkan-Loader's arg.gni file to
determine whether we should package libvulkan.so.1 (the correct name)
instead of libvulkan.so

All the executables we're packaging will use libvulkan.so.1 as the
soname, so we don't need to package or symlink libvulkan.so separately.

Original-Bug: 79706

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548602
Original-Revision: 5932e90a75ab08ee9ea4d206965012e01a2db1cc
GitOrigin-RevId: 066d072660c44585f9d10bea7243099afcea75cc
Change-Id: Ie41204e7e5cb5c13336556b942c2a6281d500407