[roll] Roll fuchsia Revert "[assembly] Generate the sparse, empty-minfs, and fastboot FVMs"
This reverts commit 4160866da8e9e078098f4816f7b15a2ef3052037.

We are reverting this, because a former commit (fxrev.dev/546832) broke
clients. This commit is not harmful, but is needed in order to revert
the former. An analysis for why the revert is needed is found in
fxrev.dev/549885.

Original-Bug: b/192085906
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549886
Original-Revision: 4c698d5d8903f56d1738a12549bdb293390e32ec
GitOrigin-RevId: 2766e311fffbc2f2906b2204f3719dac15b2c2d1
Change-Id: I1aa60b0a322d20ec271ae58165a57ba3b9fb1ace