[roll] Roll fuchsia [debugger] Improve containing function search.
Adds a flag to the CodeBlock::GetContainingFunction() function to
control whether inlines are accepted or not. This is not currently used
but will be used in followup.

To support this, allow going to the enclosing function above an inline.
Adds a new CodeBlock::GetContainingBlock() function to abstract the
confusing differences between parent() and containing_block() for the
different types of blocks.

Original-Bug: 6322
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/548943
Original-Revision: e132b1f204c0ad7211a11951cdfbf84bc589a361
GitOrigin-RevId: e289f8f9e8b4695c993dbd207bf48d81c0114b81
Change-Id: I3b23de2f09288866b2919e736ab86c6597276263