[roll] Roll fuchsia Revert "[build] Deprecate max_fvm_size: part 2"
This reverts commit 694fa050037c727ba7d4be76dffca251a33d6969.

We are reverting this, because a former commit (fxrev.dev/546832) broke
clients. This commit is not harmful, but is needed in order to revert
the former. An analysis for why the revert is needed is found in
fxrev.dev/549885.

Original-Bug: b/192085906
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549887
Original-Revision: 9bdd714ab63099d9cd562c5f6f531481d0fe78f5
GitOrigin-RevId: e7e733f4a4f76eff1fcbd88929c21bbfbd9cba2c
Change-Id: Ifbe6e4853a884f962a36023dc495a7eb16db2604