[roll] Roll fuchsia [fdio][zxio] Add a zxio_borrow operation and use for remotes and sockets
This adds an operation to "borrow" the primary handle for a zxio object
and wires it up for sockets and the remotes family. This operation is
inherently dangerous but is useful for interacting with an object in
combination with external information about how it is being used, such
as calling fdio_unsafe_borrow_channel while holding an fdio_t reference.

This also removes some knowledge of zxio object's internal knowledge
from fdio.

Original-Bug: 43267
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549888
Original-Revision: fbe0ee2f1b032535feb89f04caa21841ddd2c458
GitOrigin-RevId: ade970a2e1074b4004e70ded1a5d8bfe862ecdd2
Change-Id: I9379c2a9e6218beb63aaa90a5a70417515d71e51