[roll] Roll fuchsia [debugger] Consider CodeBlocks w/ no range empty.
Previously the CodeBlock implementation considered blocks with
unspecified code ranges as covering all instructions. As described in
the bug, this causes problems and seems contrary to the specification.
Instead, consider these empty.

In the sym-info implementation, get the FindNameContext from the frame
if there is one. This will prevent divergence in the behavior. The code
was written the way it was because the FindNameContext didn't used to be
exposed.

Original-Bug: 79797
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549403
Original-Revision: c1a285f6995d00ddd07a2af9450b95a310c8c44c
GitOrigin-RevId: 28374afc7328bbed1e0f6cb33327a6d1206a9260
Change-Id: Ic71eea9327b3c39f9a1239a082e91603e945dd74