[roll] Roll fuchsia Revert "[assembly] Do not supply fvm max size if not set"
This reverts commit 11dcd4b6bfa6cb1289e93b361e2dc26ee6f9d804.

This change broke clients, because their FVM could not expand post-boot.

The FVM has a set amount of space allocated for metadata, which directly
correlates to the maximum size the FVM can expand. The amount of metadata
is set at assembly time, and cannot change. If `--max-disk-size` is not
set at assembly time, then the FVM is assembled with a minimum space
reserved for metadata, which means that the FVM cannot expand at all.

This change removed the `--max-disk-size` flag for clients that did not
supply it explicitly, hiding the fact that clients _should_ have been
supplying it.

Original-Bug: b/192085906
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/549885
Original-Revision: c7f04acba9b8d39a656c8ccc17240684777147de
GitOrigin-RevId: 7297b995691a080e3c45de55ae5fde6ae2f9bb7e
Change-Id: I83b8bbcd33dbd9be32e69365605633fb1600a719