[roll] Roll fuchsia [driver-index] Use driver-components
This CL updates the driver-index to read a driver component
manifest in order to find where the driver library and the
bind program are located.

Once driver-index is included in the build, DriverManager will
use driver-index to find and bind packages that live in base
drivers.

The ddk-firmware-test goes through the whole process of
DriverManager interacting with DriverIndex to load a driver
from a package.

Original-Fixed: 78958
TEST: ddk-firmware-test, driver-index-unittests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/546906
Original-Revision: 42336cb5b9d6b2ca760506fabb382f1e05629b17
GitOrigin-RevId: 44b38d1caaf44c3ddeddc6d4a8b536a53de63f56
Change-Id: Ie22b0190af0c19b2ac4b56d44435bfcb66945ef5