[roll] Roll fuchsia [gidl] Use gidlir.Value type consistently
This CL replaces many uses of interface{} with gidlir.Value. The latter
is a wrapper around interface{} with the possible types it can contain
documented in value.go. It was only used in a few places before, and not
in others for no apparent reason.

Test: fx test gidl conformance
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550161
Original-Revision: 4c79702a0f3cae23776ada43780a40913d06cddb
GitOrigin-RevId: 8aa5bda6b4b0362784dc46a81456506d99746659
Change-Id: Iba1f11cbba9568058768f14029ac88e6cb480cdd