[roll] Roll fuchsia [network-device] Migrate tests to WireSharedClient
fidl::Client makes an upcall to user objects to indicate the binding has
been torn down. This behavior isn't a good fit when the user object owns
the client, since it easily leads to use-after-free of user objects.

We are splitting fidl::Client to two separate client types that are more
specialized to different threading needs, in doing so deprecating
fidl::Client.

fidl::WireSharedClient is a client type that is designed to own
the user object. By inverting the ownership relationship, it is possible
to move or clone a WireSharedClient to a different thread, with the
user object managed through a two-phase shutdown pattern. It is suitable
for multi-threaded or functional usage styles.

See //zircon/system/ulib/fidl/include/lib/fidl/llcpp/client.h for more
details.

TEST: fx test -od netdev-client-cpp-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550721
Original-Revision: 15054e15f9eb1e867a8c9087aa24b5a4dafdd3d9
GitOrigin-RevId: ef62244d5638ae709e2429c47c4e3456c32710ad
Change-Id: Ie0695bb98b7f74909c16ac441d716df6903a749a