[roll] Roll fuchsia [network-tun] Migrate to fidl::WireClient
fidl::Client makes an upcall to user objects to indicate the binding has
been torn down. This behavior isn't a good fit when the user object owns
the client, since it easily leads to use-after-free of user objects.

fidl::WireClient differs from fidl::Client in three aspects that makes
use-after-free much less likely:

- fidl::WireClient only reports errors via the `on_fidl_error` virtual
method of the event handler. User-initiated teardown (e.g. by
destroying the client) is not reported to the event handler.

- fidl::WireClient does not own the event handler. Instead, the user
object which owns the client should implement the event handling
interface, and pass a borrowed pointer to the client.

- fidl::WireClient checks that it is bound and destroyed on the
dispatcher thread through which it is using to handle FIDL messages.
This ensures that the containing user object is not destroyed while a
FIDL message or error event is being dispatched. It is suitable for
single-threaded and object oriented usage styles.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550722
Original-Revision: 2a398e7d6ed9769424ec0e5a411801b07765650f
GitOrigin-RevId: 087ab07dff8571ab74b6d1f4781888ece618591e
Change-Id: I36c886bbeac8b33ceeb3e1f85d0db7e063a00e56