[roll] Roll fuchsia [bt][a2dp] Remove "eager" designation for component
Because bt-a2dp is included in the "universe" set in some
configurations, we run into issues in system startup due to issues
described in fxbug.dev/78801. As a temporary measure, remove the "eager"
tag for the component - this will not be an issue as the component will
be started when a client requests any A2DP-providing capabilities.

Original-Bug: 79964
Test: Manifest only, fx build (this core_shard is currently not used)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/550401
Original-Revision: e37bd95eac5c1eda3890ce2274e9ba173def837b
GitOrigin-RevId: 14fdcc9e67cb8d3487c4375152e816a28fec8db7
Change-Id: I34c11d1b15826b9f9f8bec99ce4c85844b9dacfb