[roll] Roll fuchsia [bt][manifest-smoke-test] Fix flake in bt-init smoke test
Serve a pseudo-directory to bt-init (and thus bt-gap) so that bt-gap
does not panic when it tries to open the bt-host device directory.

Also, wait for realm destruction prior to ending the test, to ensure
the mock component running the pseudo-directory terminates after the
bt-gap component in the realm depending on it.

MULTIPLY: bt-init-smoke-test (fuchsia): 150
Original-Fixed: 77775
Test: First, apply fxrev.dev/541128 and verify that this patch reliably
      reproduces the flake seen in fxbug.dev/77775. Then, apply this
      patchset and verify that the flake is fixed.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/541129
Original-Revision: 29ec76816744d6a318c9bcbd178bd186394a85c3
GitOrigin-RevId: 404fe24dee5d13129b7eb5c27b44704e0342e2b2
Change-Id: Iade26bbb6edc37fe16aca831cf1289d8a8594f50